put react app in firebase + combined/renamed classes to match techspec
