+++ TechSpec.md
@@ -3,7 +3,7 @@
 ## Phaser
 - Overview: Phaser is a fast, free, and open-source HTML5 game framework for creating 2D games for desktop and mobile web browsers.
 - Features:
-- Supports both Canvas and WebGL rendering,     ensuring broad compatibility.
+- Supports both Canvas and WebGL rendering, ensuring broad compatibility.
 - Rich plugin ecosystem and a large community, providing extensive resources and support.
 - Comprehensive documentation and tutorials to assist developers at all levels.
 - Pros
