+++ public/static/css/main.ab51a433.css
@@ -1,0 +1,2 @@
+*{box-sizing:border-box;margin:0;padding:0}body{font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,sans-serif}.rooms-menu,body{background-color:#f8f7fc}.rooms-menu{align-items:center;display:flex;flex-direction:column;min-height:100vh;padding:2rem 1rem;position:relative}.room-title{color:#1a1a1a;font-size:2rem;font-weight:700;position:absolute;text-align:center;top:4rem;z-index:2}.room-display{flex-grow:1;margin-top:2rem;padding:2rem;position:relative}.room,.room-display{align-items:center;display:flex;justify-content:center}.room{aspect-ratio:16/9;background:linear-gradient(180deg,#fff,#f0f0f0);border:1px solid #ddd;border-radius:24px;box-shadow:0 8px 24px #0000001a;max-width:900px;width:85%}.overlay-buttons{display:flex;gap:.5rem;position:absolute;right:1rem;top:1rem;z-index:10}.overlay-buttons button{background-color:#6a5acd;border:none;border-radius:12px;color:#fff;cursor:pointer;font-size:.9rem;font-weight:500;padding:.5rem 1rem;transition:background-color .3s ease}.overlay-buttons button:hover{background-color:#5848b5}.media-decor-buttons{display:flex;gap:1.5rem;justify-content:center;margin:2rem 0}.media-decor-buttons button{align-items:center;background-color:#6a5acd;border:none;border-radius:50%;box-shadow:0 4px 12px #0000001a;color:#fff;cursor:pointer;display:flex;font-size:1.5rem;height:3rem;justify-content:center;transition:background-color .3s ease,transform .2s ease;width:3rem}.media-decor-buttons button:hover{background-color:#5848b5;transform:scale(1.1)}.room-navigation{align-items:center;background-color:#fff;border-top:1px solid #eee;box-shadow:0 -4px 12px #0000000d;display:flex;justify-content:space-between;margin-top:auto;padding:1rem 2rem}.rooms-list{display:flex;gap:1rem}.room-button{background:#f8f9fc;border-radius:20px;color:#333;cursor:pointer;font-size:1rem;padding:.75rem 1.5rem;transition:all .3s ease}.room-button.active{background:#6a5acd;color:#fff;font-weight:500}.room-button:hover{background:#eaeafc}.add-room-button{align-items:center;background:#6a5acd;border-radius:20px;color:#fff;cursor:pointer;display:flex;font-size:1rem;gap:.5rem;padding:.75rem 1.5rem;transition:all .3s ease}.add-room-button:hover{background:#5848b5}.modal-overlay{align-items:center;background-color:#00000080;display:flex;inset:0;justify-content:center;position:fixed;z-index:1000}.modal-content{animation:fadeIn .3s ease;background:#fff;border-radius:24px;box-shadow:0 8px 24px #0003;max-width:400px;padding:2rem;width:90%}@keyframes fadeIn{0%{opacity:0;transform:translateY(-10%)}to{opacity:1;transform:translateY(0)}}.top-menu-bar{align-items:center;background-color:#fff;box-shadow:0 2px 6px #0000001a;display:flex;height:3rem;justify-content:space-between;left:0;padding:0 1rem;position:fixed;right:0;top:0;z-index:10}.house-name{color:#1a1a1a;cursor:default;font-size:1.5rem;font-weight:700;margin-right:auto;transition:color .3s ease}.house-name:hover{color:#6a5acd}.settings-button{background:none;border:none;color:#6a5acd;cursor:pointer;font-size:1.25rem;transition:color .3s ease}.settings-button:hover{color:#5848b5}
+/*# sourceMappingURL=main.ab51a433.css.map*/
