+++ public/static/js/main.f9304b0b.js
@@ -1,0 +1,3 @@
+/*! For license information please see main.f9304b0b.js.LICENSE.txt */
+(()=>{"use strict";var e={730:(e,n,t)=>{var r=t(43),l=t(853);function a(e){for(var n="https://reactjs.org/docs/error-decoder.html?invariant="+e,t=1;t<arguments.length;t++)n+="&args[]="+encodeURIComponent(arguments[t]);return"Minified React error #"+e+"; visit "+n+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var o=new Set,u={};function i(e,n){s(e,n),s(e+"Capture",n)}function s(e,n){for(u[e]=n,e=0;e<n.length;e++)o.add(n[e])}var c=!("undefined"===typeof window||"undefined"===typeof window.document||"undefined"===typeof window.document.createElement),f=Object.prototype.hasOwnProperty,d=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,p={},m={};function h(e,n,t,r,l,a,o){this.acceptsBooleans=2===n||3===n||4===n,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=t,this.propertyName=e,this.type=n,this.sanitizeURL=a,this.removeEmptyString=o}var v={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach((function(e){v[e]=new h(e,0,!1,e,null,!1,!1)})),[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach((function(e){var n=e[0];v[n]=new h(n,1,!1,e[1],null,!1,!1)})),["contentEditable","draggable","spellCheck","value"].forEach((function(e){v[e]=new h(e,2,!1,e.toLowerCase(),null,!1,!1)})),["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach((function(e){v[e]=new h(e,2,!1,e,null,!1,!1)})),"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach((function(e){v[e]=new h(e,3,!1,e.toLowerCase(),null,!1,!1)})),["checked","multiple","muted","selected"].forEach((function(e){v[e]=new h(e,3,!0,e,null,!1,!1)})),["capture","download"].forEach((function(e){v[e]=new h(e,4,!1,e,null,!1,!1)})),["cols","rows","size","span"].forEach((function(e){v[e]=new h(e,6,!1,e,null,!1,!1)})),["rowSpan","start"].forEach((function(e){v[e]=new h(e,5,!1,e.toLowerCase(),null,!1,!1)}));var g=/[\-:]([a-z])/g;function y(e){return e[1].toUpperCase()}function b(e,n,t,r){var l=v.hasOwnProperty(n)?v[n]:null;(null!==l?0!==l.type:r||!(2<n.length)||"o"!==n[0]&&"O"!==n[0]||"n"!==n[1]&&"N"!==n[1])&&(function(e,n,t,r){if(null===n||"undefined"===typeof n||function(e,n,t,r){if(null!==t&&0===t.type)return!1;switch(typeof n){case"function":case"symbol":return!0;case"boolean":return!r&&(null!==t?!t.acceptsBooleans:"data-"!==(e=e.toLowerCase().slice(0,5))&&"aria-"!==e);default:return!1}}(e,n,t,r))return!0;if(r)return!1;if(null!==t)switch(t.type){case 3:return!n;case 4:return!1===n;case 5:return isNaN(n);case 6:return isNaN(n)||1>n}return!1}(n,t,l,r)&&(t=null),r||null===l?function(e){return!!f.call(m,e)||!f.call(p,e)&&(d.test(e)?m[e]=!0:(p[e]=!0,!1))}(n)&&(null===t?e.removeAttribute(n):e.setAttribute(n,""+t)):l.mustUseProperty?e[l.propertyName]=null===t?3!==l.type&&"":t:(n=l.attributeName,r=l.attributeNamespace,null===t?e.removeAttribute(n):(t=3===(l=l.type)||4===l&&!0===t?"":""+t,r?e.setAttributeNS(r,n,t):e.setAttribute(n,t))))}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach((function(e){var n=e.replace(g,y);v[n]=new h(n,1,!1,e,null,!1,!1)})),"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach((function(e){var n=e.replace(g,y);v[n]=new h(n,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)})),["xml:base","xml:lang","xml:space"].forEach((function(e){var n=e.replace(g,y);v[n]=new h(n,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)})),["tabIndex","crossOrigin"].forEach((function(e){v[e]=new h(e,1,!1,e.toLowerCase(),null,!1,!1)})),v.xlinkHref=new h("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1),["src","href","action","formAction"].forEach((function(e){v[e]=new h(e,1,!1,e.toLowerCase(),null,!0,!0)}));var k=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,w=Symbol.for("react.element"),S=Symbol.for("react.portal"),x=Symbol.for("react.fragment"),E=Symbol.for("react.strict_mode"),C=Symbol.for("react.profiler"),_=Symbol.for("react.provider"),N=Symbol.for("react.context"),P=Symbol.for("react.forward_ref"),z=Symbol.for("react.suspense"),L=Symbol.for("react.suspense_list"),T=Symbol.for("react.memo"),R=Symbol.for("react.lazy");Symbol.for("react.scope"),Symbol.for("react.debug_trace_mode");var O=Symbol.for("react.offscreen");Symbol.for("react.legacy_hidden"),Symbol.for("react.cache"),Symbol.for("react.tracing_marker");var M=Symbol.iterator;function j(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=M&&e[M]||e["@@iterator"])?e:null}var F,I=Object.assign;function D(e){if(void 0===F)try{throw Error()}catch(t){var n=t.stack.trim().match(/\n( *(at )?)/);F=n&&n[1]||""}return"\n"+F+e}var U=!1;function $(e,n){if(!e||U)return"";U=!0;var t=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(n)if(n=function(){throw Error()},Object.defineProperty(n.prototype,"props",{set:function(){throw Error()}}),"object"===typeof Reflect&&Reflect.construct){try{Reflect.construct(n,[])}catch(s){var r=s}Reflect.construct(e,[],n)}else{try{n.call()}catch(s){r=s}e.call(n.prototype)}else{try{throw Error()}catch(s){r=s}e()}}catch(s){if(s&&r&&"string"===typeof s.stack){for(var l=s.stack.split("\n"),a=r.stack.split("\n"),o=l.length-1,u=a.length-1;1<=o&&0<=u&&l[o]!==a[u];)u--;for(;1<=o&&0<=u;o--,u--)if(l[o]!==a[u]){if(1!==o||1!==u)do{if(o--,0>--u||l[o]!==a[u]){var i="\n"+l[o].replace(" at new "," at ");return e.displayName&&i.includes("<anonymous>")&&(i=i.replace("<anonymous>",e.displayName)),i}}while(1<=o&&0<=u);break}}}finally{U=!1,Error.prepareStackTrace=t}return(e=e?e.displayName||e.name:"")?D(e):""}function A(e){switch(e.tag){case 5:return D(e.type);case 16:return D("Lazy");case 13:return D("Suspense");case 19:return D("SuspenseList");case 0:case 2:case 15:return e=$(e.type,!1);case 11:return e=$(e.type.render,!1);case 1:return e=$(e.type,!0);default:return""}}function H(e){if(null==e)return null;if("function"===typeof e)return e.displayName||e.name||null;if("string"===typeof e)return e;switch(e){case x:return"Fragment";case S:return"Portal";case C:return"Profiler";case E:return"StrictMode";case z:return"Suspense";case L:return"SuspenseList"}if("object"===typeof e)switch(e.$$typeof){case N:return(e.displayName||"Context")+".Consumer";case _:return(e._context.displayName||"Context")+".Provider";case P:var n=e.render;return(e=e.displayName)||(e=""!==(e=n.displayName||n.name||"")?"ForwardRef("+e+")":"ForwardRef"),e;case T:return null!==(n=e.displayName||null)?n:H(e.type)||"Memo";case R:n=e._payload,e=e._init;try{return H(e(n))}catch(t){}}return null}function V(e){var n=e.type;switch(e.tag){case 24:return"Cache";case 9:return(n.displayName||"Context")+".Consumer";case 10:return(n._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=(e=n.render).displayName||e.name||"",n.displayName||(""!==e?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return n;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return H(n);case 8:return n===E?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if("function"===typeof n)return n.displayName||n.name||null;if("string"===typeof n)return n}return null}function B(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":case"object":return e;default:return""}}function W(e){var n=e.type;return(e=e.nodeName)&&"input"===e.toLowerCase()&&("checkbox"===n||"radio"===n)}function Q(e){e._valueTracker||(e._valueTracker=function(e){var n=W(e)?"checked":"value",t=Object.getOwnPropertyDescriptor(e.constructor.prototype,n),r=""+e[n];if(!e.hasOwnProperty(n)&&"undefined"!==typeof t&&"function"===typeof t.get&&"function"===typeof t.set){var l=t.get,a=t.set;return Object.defineProperty(e,n,{configurable:!0,get:function(){return l.call(this)},set:function(e){r=""+e,a.call(this,e)}}),Object.defineProperty(e,n,{enumerable:t.enumerable}),{getValue:function(){return r},setValue:function(e){r=""+e},stopTracking:function(){e._valueTracker=null,delete e[n]}}}}(e))}function q(e){if(!e)return!1;var n=e._valueTracker;if(!n)return!0;var t=n.getValue(),r="";return e&&(r=W(e)?e.checked?"true":"false":e.value),(e=r)!==t&&(n.setValue(e),!0)}function K(e){if("undefined"===typeof(e=e||("undefined"!==typeof document?document:void 0)))return null;try{return e.activeElement||e.body}catch(n){return e.body}}function Y(e,n){var t=n.checked;return I({},n,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:null!=t?t:e._wrapperState.initialChecked})}function X(e,n){var t=null==n.defaultValue?"":n.defaultValue,r=null!=n.checked?n.checked:n.defaultChecked;t=B(null!=n.value?n.value:t),e._wrapperState={initialChecked:r,initialValue:t,controlled:"checkbox"===n.type||"radio"===n.type?null!=n.checked:null!=n.value}}function G(e,n){null!=(n=n.checked)&&b(e,"checked",n,!1)}function J(e,n){G(e,n);var t=B(n.value),r=n.type;if(null!=t)"number"===r?(0===t&&""===e.value||e.value!=t)&&(e.value=""+t):e.value!==""+t&&(e.value=""+t);else if("submit"===r||"reset"===r)return void e.removeAttribute("value");n.hasOwnProperty("value")?ee(e,n.type,t):n.hasOwnProperty("defaultValue")&&ee(e,n.type,B(n.defaultValue)),null==n.checked&&null!=n.defaultChecked&&(e.defaultChecked=!!n.defaultChecked)}function Z(e,n,t){if(n.hasOwnProperty("value")||n.hasOwnProperty("defaultValue")){var r=n.type;if(!("submit"!==r&&"reset"!==r||void 0!==n.value&&null!==n.value))return;n=""+e._wrapperState.initialValue,t||n===e.value||(e.value=n),e.defaultValue=n}""!==(t=e.name)&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,""!==t&&(e.name=t)}function ee(e,n,t){"number"===n&&K(e.ownerDocument)===e||(null==t?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+t&&(e.defaultValue=""+t))}var ne=Array.isArray;function te(e,n,t,r){if(e=e.options,n){n={};for(var l=0;l<t.length;l++)n["$"+t[l]]=!0;for(t=0;t<e.length;t++)l=n.hasOwnProperty("$"+e[t].value),e[t].selected!==l&&(e[t].selected=l),l&&r&&(e[t].defaultSelected=!0)}else{for(t=""+B(t),n=null,l=0;l<e.length;l++){if(e[l].value===t)return e[l].selected=!0,void(r&&(e[l].defaultSelected=!0));null!==n||e[l].disabled||(n=e[l])}null!==n&&(n.selected=!0)}}function re(e,n){if(null!=n.dangerouslySetInnerHTML)throw Error(a(91));return I({},n,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function le(e,n){var t=n.value;if(null==t){if(t=n.children,n=n.defaultValue,null!=t){if(null!=n)throw Error(a(92));if(ne(t)){if(1<t.length)throw Error(a(93));t=t[0]}n=t}null==n&&(n=""),t=n}e._wrapperState={initialValue:B(t)}}function ae(e,n){var t=B(n.value),r=B(n.defaultValue);null!=t&&((t=""+t)!==e.value&&(e.value=t),null==n.defaultValue&&e.defaultValue!==t&&(e.defaultValue=t)),null!=r&&(e.defaultValue=""+r)}function oe(e){var n=e.textContent;n===e._wrapperState.initialValue&&""!==n&&null!==n&&(e.value=n)}function ue(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function ie(e,n){return null==e||"http://www.w3.org/1999/xhtml"===e?ue(n):"http://www.w3.org/2000/svg"===e&&"foreignObject"===n?"http://www.w3.org/1999/xhtml":e}var se,ce,fe=(ce=function(e,n){if("http://www.w3.org/2000/svg"!==e.namespaceURI||"innerHTML"in e)e.innerHTML=n;else{for((se=se||document.createElement("div")).innerHTML="<svg>"+n.valueOf().toString()+"</svg>",n=se.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;n.firstChild;)e.appendChild(n.firstChild)}},"undefined"!==typeof MSApp&&MSApp.execUnsafeLocalFunction?function(e,n,t,r){MSApp.execUnsafeLocalFunction((function(){return ce(e,n)}))}:ce);function de(e,n){if(n){var t=e.firstChild;if(t&&t===e.lastChild&&3===t.nodeType)return void(t.nodeValue=n)}e.textContent=n}var pe={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},me=["Webkit","ms","Moz","O"];function he(e,n,t){return null==n||"boolean"===typeof n||""===n?"":t||"number"!==typeof n||0===n||pe.hasOwnProperty(e)&&pe[e]?(""+n).trim():n+"px"}function ve(e,n){for(var t in e=e.style,n)if(n.hasOwnProperty(t)){var r=0===t.indexOf("--"),l=he(t,n[t],r);"float"===t&&(t="cssFloat"),r?e.setProperty(t,l):e[t]=l}}Object.keys(pe).forEach((function(e){me.forEach((function(n){n=n+e.charAt(0).toUpperCase()+e.substring(1),pe[n]=pe[e]}))}));var ge=I({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function ye(e,n){if(n){if(ge[e]&&(null!=n.children||null!=n.dangerouslySetInnerHTML))throw Error(a(137,e));if(null!=n.dangerouslySetInnerHTML){if(null!=n.children)throw Error(a(60));if("object"!==typeof n.dangerouslySetInnerHTML||!("__html"in n.dangerouslySetInnerHTML))throw Error(a(61))}if(null!=n.style&&"object"!==typeof n.style)throw Error(a(62))}}function be(e,n){if(-1===e.indexOf("-"))return"string"===typeof n.is;switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var ke=null;function we(e){return(e=e.target||e.srcElement||window).correspondingUseElement&&(e=e.correspondingUseElement),3===e.nodeType?e.parentNode:e}var Se=null,xe=null,Ee=null;function Ce(e){if(e=bl(e)){if("function"!==typeof Se)throw Error(a(280));var n=e.stateNode;n&&(n=wl(n),Se(e.stateNode,e.type,n))}}function _e(e){xe?Ee?Ee.push(e):Ee=[e]:xe=e}function Ne(){if(xe){var e=xe,n=Ee;if(Ee=xe=null,Ce(e),n)for(e=0;e<n.length;e++)Ce(n[e])}}function Pe(e,n){return e(n)}function ze(){}var Le=!1;function Te(e,n,t){if(Le)return e(n,t);Le=!0;try{return Pe(e,n,t)}finally{Le=!1,(null!==xe||null!==Ee)&&(ze(),Ne())}}function Re(e,n){var t=e.stateNode;if(null===t)return null;var r=wl(t);if(null===r)return null;t=r[n];e:switch(n){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(r=!("button"===(e=e.type)||"input"===e||"select"===e||"textarea"===e)),e=!r;break e;default:e=!1}if(e)return null;if(t&&"function"!==typeof t)throw Error(a(231,n,typeof t));return t}var Oe=!1;if(c)try{var Me={};Object.defineProperty(Me,"passive",{get:function(){Oe=!0}}),window.addEventListener("test",Me,Me),window.removeEventListener("test",Me,Me)}catch(ce){Oe=!1}function je(e,n,t,r,l,a,o,u,i){var s=Array.prototype.slice.call(arguments,3);try{n.apply(t,s)}catch(c){this.onError(c)}}var Fe=!1,Ie=null,De=!1,Ue=null,$e={onError:function(e){Fe=!0,Ie=e}};function Ae(e,n,t,r,l,a,o,u,i){Fe=!1,Ie=null,je.apply($e,arguments)}function He(e){var n=e,t=e;if(e.alternate)for(;n.return;)n=n.return;else{e=n;do{0!==(4098&(n=e).flags)&&(t=n.return),e=n.return}while(e)}return 3===n.tag?t:null}function Ve(e){if(13===e.tag){var n=e.memoizedState;if(null===n&&(null!==(e=e.alternate)&&(n=e.memoizedState)),null!==n)return n.dehydrated}return null}function Be(e){if(He(e)!==e)throw Error(a(188))}function We(e){return null!==(e=function(e){var n=e.alternate;if(!n){if(null===(n=He(e)))throw Error(a(188));return n!==e?null:e}for(var t=e,r=n;;){var l=t.return;if(null===l)break;var o=l.alternate;if(null===o){if(null!==(r=l.return)){t=r;continue}break}if(l.child===o.child){for(o=l.child;o;){if(o===t)return Be(l),e;if(o===r)return Be(l),n;o=o.sibling}throw Error(a(188))}if(t.return!==r.return)t=l,r=o;else{for(var u=!1,i=l.child;i;){if(i===t){u=!0,t=l,r=o;break}if(i===r){u=!0,r=l,t=o;break}i=i.sibling}if(!u){for(i=o.child;i;){if(i===t){u=!0,t=o,r=l;break}if(i===r){u=!0,r=o,t=l;break}i=i.sibling}if(!u)throw Error(a(189))}}if(t.alternate!==r)throw Error(a(190))}if(3!==t.tag)throw Error(a(188));return t.stateNode.current===t?e:n}(e))?Qe(e):null}function Qe(e){if(5===e.tag||6===e.tag)return e;for(e=e.child;null!==e;){var n=Qe(e);if(null!==n)return n;e=e.sibling}return null}var qe=l.unstable_scheduleCallback,Ke=l.unstable_cancelCallback,Ye=l.unstable_shouldYield,Xe=l.unstable_requestPaint,Ge=l.unstable_now,Je=l.unstable_getCurrentPriorityLevel,Ze=l.unstable_ImmediatePriority,en=l.unstable_UserBlockingPriority,nn=l.unstable_NormalPriority,tn=l.unstable_LowPriority,rn=l.unstable_IdlePriority,ln=null,an=null;var on=Math.clz32?Math.clz32:function(e){return e>>>=0,0===e?32:31-(un(e)/sn|0)|0},un=Math.log,sn=Math.LN2;var cn=64,fn=4194304;function dn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return 4194240&e;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return 130023424&e;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function pn(e,n){var t=e.pendingLanes;if(0===t)return 0;var r=0,l=e.suspendedLanes,a=e.pingedLanes,o=268435455&t;if(0!==o){var u=o&~l;0!==u?r=dn(u):0!==(a&=o)&&(r=dn(a))}else 0!==(o=t&~l)?r=dn(o):0!==a&&(r=dn(a));if(0===r)return 0;if(0!==n&&n!==r&&0===(n&l)&&((l=r&-r)>=(a=n&-n)||16===l&&0!==(4194240&a)))return n;if(0!==(4&r)&&(r|=16&t),0!==(n=e.entangledLanes))for(e=e.entanglements,n&=r;0<n;)l=1<<(t=31-on(n)),r|=e[t],n&=~l;return r}function mn(e,n){switch(e){case 1:case 2:case 4:return n+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return n+5e3;default:return-1}}function hn(e){return 0!==(e=-1073741825&e.pendingLanes)?e:1073741824&e?1073741824:0}function vn(){var e=cn;return 0===(4194240&(cn<<=1))&&(cn=64),e}function gn(e){for(var n=[],t=0;31>t;t++)n.push(e);return n}function yn(e,n,t){e.pendingLanes|=n,536870912!==n&&(e.suspendedLanes=0,e.pingedLanes=0),(e=e.eventTimes)[n=31-on(n)]=t}function bn(e,n){var t=e.entangledLanes|=n;for(e=e.entanglements;t;){var r=31-on(t),l=1<<r;l&n|e[r]&n&&(e[r]|=n),t&=~l}}var kn=0;function wn(e){return 1<(e&=-e)?4<e?0!==(268435455&e)?16:536870912:4:1}var Sn,xn,En,Cn,_n,Nn=!1,Pn=[],zn=null,Ln=null,Tn=null,Rn=new Map,On=new Map,Mn=[],jn="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function Fn(e,n){switch(e){case"focusin":case"focusout":zn=null;break;case"dragenter":case"dragleave":Ln=null;break;case"mouseover":case"mouseout":Tn=null;break;case"pointerover":case"pointerout":Rn.delete(n.pointerId);break;case"gotpointercapture":case"lostpointercapture":On.delete(n.pointerId)}}function In(e,n,t,r,l,a){return null===e||e.nativeEvent!==a?(e={blockedOn:n,domEventName:t,eventSystemFlags:r,nativeEvent:a,targetContainers:[l]},null!==n&&(null!==(n=bl(n))&&xn(n)),e):(e.eventSystemFlags|=r,n=e.targetContainers,null!==l&&-1===n.indexOf(l)&&n.push(l),e)}function Dn(e){var n=yl(e.target);if(null!==n){var t=He(n);if(null!==t)if(13===(n=t.tag)){if(null!==(n=Ve(t)))return e.blockedOn=n,void _n(e.priority,(function(){En(t)}))}else if(3===n&&t.stateNode.current.memoizedState.isDehydrated)return void(e.blockedOn=3===t.tag?t.stateNode.containerInfo:null)}e.blockedOn=null}function Un(e){if(null!==e.blockedOn)return!1;for(var n=e.targetContainers;0<n.length;){var t=Xn(e.domEventName,e.eventSystemFlags,n[0],e.nativeEvent);if(null!==t)return null!==(n=bl(t))&&xn(n),e.blockedOn=t,!1;var r=new(t=e.nativeEvent).constructor(t.type,t);ke=r,t.target.dispatchEvent(r),ke=null,n.shift()}return!0}function $n(e,n,t){Un(e)&&t.delete(n)}function An(){Nn=!1,null!==zn&&Un(zn)&&(zn=null),null!==Ln&&Un(Ln)&&(Ln=null),null!==Tn&&Un(Tn)&&(Tn=null),Rn.forEach($n),On.forEach($n)}function Hn(e,n){e.blockedOn===n&&(e.blockedOn=null,Nn||(Nn=!0,l.unstable_scheduleCallback(l.unstable_NormalPriority,An)))}function Vn(e){function n(n){return Hn(n,e)}if(0<Pn.length){Hn(Pn[0],e);for(var t=1;t<Pn.length;t++){var r=Pn[t];r.blockedOn===e&&(r.blockedOn=null)}}for(null!==zn&&Hn(zn,e),null!==Ln&&Hn(Ln,e),null!==Tn&&Hn(Tn,e),Rn.forEach(n),On.forEach(n),t=0;t<Mn.length;t++)(r=Mn[t]).blockedOn===e&&(r.blockedOn=null);for(;0<Mn.length&&null===(t=Mn[0]).blockedOn;)Dn(t),null===t.blockedOn&&Mn.shift()}var Bn=k.ReactCurrentBatchConfig,Wn=!0;function Qn(e,n,t,r){var l=kn,a=Bn.transition;Bn.transition=null;try{kn=1,Kn(e,n,t,r)}finally{kn=l,Bn.transition=a}}function qn(e,n,t,r){var l=kn,a=Bn.transition;Bn.transition=null;try{kn=4,Kn(e,n,t,r)}finally{kn=l,Bn.transition=a}}function Kn(e,n,t,r){if(Wn){var l=Xn(e,n,t,r);if(null===l)Br(e,n,r,Yn,t),Fn(e,r);else if(function(e,n,t,r,l){switch(n){case"focusin":return zn=In(zn,e,n,t,r,l),!0;case"dragenter":return Ln=In(Ln,e,n,t,r,l),!0;case"mouseover":return Tn=In(Tn,e,n,t,r,l),!0;case"pointerover":var a=l.pointerId;return Rn.set(a,In(Rn.get(a)||null,e,n,t,r,l)),!0;case"gotpointercapture":return a=l.pointerId,On.set(a,In(On.get(a)||null,e,n,t,r,l)),!0}return!1}(l,e,n,t,r))r.stopPropagation();else if(Fn(e,r),4&n&&-1<jn.indexOf(e)){for(;null!==l;){var a=bl(l);if(null!==a&&Sn(a),null===(a=Xn(e,n,t,r))&&Br(e,n,r,Yn,t),a===l)break;l=a}null!==l&&r.stopPropagation()}else Br(e,n,r,null,t)}}var Yn=null;function Xn(e,n,t,r){if(Yn=null,null!==(e=yl(e=we(r))))if(null===(n=He(e)))e=null;else if(13===(t=n.tag)){if(null!==(e=Ve(n)))return e;e=null}else if(3===t){if(n.stateNode.current.memoizedState.isDehydrated)return 3===n.tag?n.stateNode.containerInfo:null;e=null}else n!==e&&(e=null);return Yn=e,null}function Gn(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Je()){case Ze:return 1;case en:return 4;case nn:case tn:return 16;case rn:return 536870912;default:return 16}default:return 16}}var Jn=null,Zn=null,et=null;function nt(){if(et)return et;var e,n,t=Zn,r=t.length,l="value"in Jn?Jn.value:Jn.textContent,a=l.length;for(e=0;e<r&&t[e]===l[e];e++);var o=r-e;for(n=1;n<=o&&t[r-n]===l[a-n];n++);return et=l.slice(e,1<n?1-n:void 0)}function tt(e){var n=e.keyCode;return"charCode"in e?0===(e=e.charCode)&&13===n&&(e=13):e=n,10===e&&(e=13),32<=e||13===e?e:0}function rt(){return!0}function lt(){return!1}function at(e){function n(n,t,r,l,a){for(var o in this._reactName=n,this._targetInst=r,this.type=t,this.nativeEvent=l,this.target=a,this.currentTarget=null,e)e.hasOwnProperty(o)&&(n=e[o],this[o]=n?n(l):l[o]);return this.isDefaultPrevented=(null!=l.defaultPrevented?l.defaultPrevented:!1===l.returnValue)?rt:lt,this.isPropagationStopped=lt,this}return I(n.prototype,{preventDefault:function(){this.defaultPrevented=!0;var e=this.nativeEvent;e&&(e.preventDefault?e.preventDefault():"unknown"!==typeof e.returnValue&&(e.returnValue=!1),this.isDefaultPrevented=rt)},stopPropagation:function(){var e=this.nativeEvent;e&&(e.stopPropagation?e.stopPropagation():"unknown"!==typeof e.cancelBubble&&(e.cancelBubble=!0),this.isPropagationStopped=rt)},persist:function(){},isPersistent:rt}),n}var ot,ut,it,st={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},ct=at(st),ft=I({},st,{view:0,detail:0}),dt=at(ft),pt=I({},ft,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:Ct,button:0,buttons:0,relatedTarget:function(e){return void 0===e.relatedTarget?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==it&&(it&&"mousemove"===e.type?(ot=e.screenX-it.screenX,ut=e.screenY-it.screenY):ut=ot=0,it=e),ot)},movementY:function(e){return"movementY"in e?e.movementY:ut}}),mt=at(pt),ht=at(I({},pt,{dataTransfer:0})),vt=at(I({},ft,{relatedTarget:0})),gt=at(I({},st,{animationName:0,elapsedTime:0,pseudoElement:0})),yt=I({},st,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),bt=at(yt),kt=at(I({},st,{data:0})),wt={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},St={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},xt={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Et(e){var n=this.nativeEvent;return n.getModifierState?n.getModifierState(e):!!(e=xt[e])&&!!n[e]}function Ct(){return Et}var _t=I({},ft,{key:function(e){if(e.key){var n=wt[e.key]||e.key;if("Unidentified"!==n)return n}return"keypress"===e.type?13===(e=tt(e))?"Enter":String.fromCharCode(e):"keydown"===e.type||"keyup"===e.type?St[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:Ct,charCode:function(e){return"keypress"===e.type?tt(e):0},keyCode:function(e){return"keydown"===e.type||"keyup"===e.type?e.keyCode:0},which:function(e){return"keypress"===e.type?tt(e):"keydown"===e.type||"keyup"===e.type?e.keyCode:0}}),Nt=at(_t),Pt=at(I({},pt,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0})),zt=at(I({},ft,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:Ct})),Lt=at(I({},st,{propertyName:0,elapsedTime:0,pseudoElement:0})),Tt=I({},pt,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),Rt=at(Tt),Ot=[9,13,27,32],Mt=c&&"CompositionEvent"in window,jt=null;c&&"documentMode"in document&&(jt=document.documentMode);var Ft=c&&"TextEvent"in window&&!jt,It=c&&(!Mt||jt&&8<jt&&11>=jt),Dt=String.fromCharCode(32),Ut=!1;function $t(e,n){switch(e){case"keyup":return-1!==Ot.indexOf(n.keyCode);case"keydown":return 229!==n.keyCode;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function At(e){return"object"===typeof(e=e.detail)&&"data"in e?e.data:null}var Ht=!1;var Vt={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Bt(e){var n=e&&e.nodeName&&e.nodeName.toLowerCase();return"input"===n?!!Vt[e.type]:"textarea"===n}function Wt(e,n,t,r){_e(r),0<(n=Qr(n,"onChange")).length&&(t=new ct("onChange","change",null,t,r),e.push({event:t,listeners:n}))}var Qt=null,qt=null;function Kt(e){Dr(e,0)}function Yt(e){if(q(kl(e)))return e}function Xt(e,n){if("change"===e)return n}var Gt=!1;if(c){var Jt;if(c){var Zt="oninput"in document;if(!Zt){var er=document.createElement("div");er.setAttribute("oninput","return;"),Zt="function"===typeof er.oninput}Jt=Zt}else Jt=!1;Gt=Jt&&(!document.documentMode||9<document.documentMode)}function nr(){Qt&&(Qt.detachEvent("onpropertychange",tr),qt=Qt=null)}function tr(e){if("value"===e.propertyName&&Yt(qt)){var n=[];Wt(n,qt,e,we(e)),Te(Kt,n)}}function rr(e,n,t){"focusin"===e?(nr(),qt=t,(Qt=n).attachEvent("onpropertychange",tr)):"focusout"===e&&nr()}function lr(e){if("selectionchange"===e||"keyup"===e||"keydown"===e)return Yt(qt)}function ar(e,n){if("click"===e)return Yt(n)}function or(e,n){if("input"===e||"change"===e)return Yt(n)}var ur="function"===typeof Object.is?Object.is:function(e,n){return e===n&&(0!==e||1/e===1/n)||e!==e&&n!==n};function ir(e,n){if(ur(e,n))return!0;if("object"!==typeof e||null===e||"object"!==typeof n||null===n)return!1;var t=Object.keys(e),r=Object.keys(n);if(t.length!==r.length)return!1;for(r=0;r<t.length;r++){var l=t[r];if(!f.call(n,l)||!ur(e[l],n[l]))return!1}return!0}function sr(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function cr(e,n){var t,r=sr(e);for(e=0;r;){if(3===r.nodeType){if(t=e+r.textContent.length,e<=n&&t>=n)return{node:r,offset:n-e};e=t}e:{for(;r;){if(r.nextSibling){r=r.nextSibling;break e}r=r.parentNode}r=void 0}r=sr(r)}}function fr(e,n){return!(!e||!n)&&(e===n||(!e||3!==e.nodeType)&&(n&&3===n.nodeType?fr(e,n.parentNode):"contains"in e?e.contains(n):!!e.compareDocumentPosition&&!!(16&e.compareDocumentPosition(n))))}function dr(){for(var e=window,n=K();n instanceof e.HTMLIFrameElement;){try{var t="string"===typeof n.contentWindow.location.href}catch(r){t=!1}if(!t)break;n=K((e=n.contentWindow).document)}return n}function pr(e){var n=e&&e.nodeName&&e.nodeName.toLowerCase();return n&&("input"===n&&("text"===e.type||"search"===e.type||"tel"===e.type||"url"===e.type||"password"===e.type)||"textarea"===n||"true"===e.contentEditable)}function mr(e){var n=dr(),t=e.focusedElem,r=e.selectionRange;if(n!==t&&t&&t.ownerDocument&&fr(t.ownerDocument.documentElement,t)){if(null!==r&&pr(t))if(n=r.start,void 0===(e=r.end)&&(e=n),"selectionStart"in t)t.selectionStart=n,t.selectionEnd=Math.min(e,t.value.length);else if((e=(n=t.ownerDocument||document)&&n.defaultView||window).getSelection){e=e.getSelection();var l=t.textContent.length,a=Math.min(r.start,l);r=void 0===r.end?a:Math.min(r.end,l),!e.extend&&a>r&&(l=r,r=a,a=l),l=cr(t,a);var o=cr(t,r);l&&o&&(1!==e.rangeCount||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==o.node||e.focusOffset!==o.offset)&&((n=n.createRange()).setStart(l.node,l.offset),e.removeAllRanges(),a>r?(e.addRange(n),e.extend(o.node,o.offset)):(n.setEnd(o.node,o.offset),e.addRange(n)))}for(n=[],e=t;e=e.parentNode;)1===e.nodeType&&n.push({element:e,left:e.scrollLeft,top:e.scrollTop});for("function"===typeof t.focus&&t.focus(),t=0;t<n.length;t++)(e=n[t]).element.scrollLeft=e.left,e.element.scrollTop=e.top}}var hr=c&&"documentMode"in document&&11>=document.documentMode,vr=null,gr=null,yr=null,br=!1;function kr(e,n,t){var r=t.window===t?t.document:9===t.nodeType?t:t.ownerDocument;br||null==vr||vr!==K(r)||("selectionStart"in(r=vr)&&pr(r)?r={start:r.selectionStart,end:r.selectionEnd}:r={anchorNode:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection()).anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset},yr&&ir(yr,r)||(yr=r,0<(r=Qr(gr,"onSelect")).length&&(n=new ct("onSelect","select",null,n,t),e.push({event:n,listeners:r}),n.target=vr)))}function wr(e,n){var t={};return t[e.toLowerCase()]=n.toLowerCase(),t["Webkit"+e]="webkit"+n,t["Moz"+e]="moz"+n,t}var Sr={animationend:wr("Animation","AnimationEnd"),animationiteration:wr("Animation","AnimationIteration"),animationstart:wr("Animation","AnimationStart"),transitionend:wr("Transition","TransitionEnd")},xr={},Er={};function Cr(e){if(xr[e])return xr[e];if(!Sr[e])return e;var n,t=Sr[e];for(n in t)if(t.hasOwnProperty(n)&&n in Er)return xr[e]=t[n];return e}c&&(Er=document.createElement("div").style,"AnimationEvent"in window||(delete Sr.animationend.animation,delete Sr.animationiteration.animation,delete Sr.animationstart.animation),"TransitionEvent"in window||delete Sr.transitionend.transition);var _r=Cr("animationend"),Nr=Cr("animationiteration"),Pr=Cr("animationstart"),zr=Cr("transitionend"),Lr=new Map,Tr="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function Rr(e,n){Lr.set(e,n),i(n,[e])}for(var Or=0;Or<Tr.length;Or++){var Mr=Tr[Or];Rr(Mr.toLowerCase(),"on"+(Mr[0].toUpperCase()+Mr.slice(1)))}Rr(_r,"onAnimationEnd"),Rr(Nr,"onAnimationIteration"),Rr(Pr,"onAnimationStart"),Rr("dblclick","onDoubleClick"),Rr("focusin","onFocus"),Rr("focusout","onBlur"),Rr(zr,"onTransitionEnd"),s("onMouseEnter",["mouseout","mouseover"]),s("onMouseLeave",["mouseout","mouseover"]),s("onPointerEnter",["pointerout","pointerover"]),s("onPointerLeave",["pointerout","pointerover"]),i("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),i("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),i("onBeforeInput",["compositionend","keypress","textInput","paste"]),i("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),i("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),i("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var jr="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Fr=new Set("cancel close invalid load scroll toggle".split(" ").concat(jr));function Ir(e,n,t){var r=e.type||"unknown-event";e.currentTarget=t,function(e,n,t,r,l,o,u,i,s){if(Ae.apply(this,arguments),Fe){if(!Fe)throw Error(a(198));var c=Ie;Fe=!1,Ie=null,De||(De=!0,Ue=c)}}(r,n,void 0,e),e.currentTarget=null}function Dr(e,n){n=0!==(4&n);for(var t=0;t<e.length;t++){var r=e[t],l=r.event;r=r.listeners;e:{var a=void 0;if(n)for(var o=r.length-1;0<=o;o--){var u=r[o],i=u.instance,s=u.currentTarget;if(u=u.listener,i!==a&&l.isPropagationStopped())break e;Ir(l,u,s),a=i}else for(o=0;o<r.length;o++){if(i=(u=r[o]).instance,s=u.currentTarget,u=u.listener,i!==a&&l.isPropagationStopped())break e;Ir(l,u,s),a=i}}}if(De)throw e=Ue,De=!1,Ue=null,e}function Ur(e,n){var t=n[hl];void 0===t&&(t=n[hl]=new Set);var r=e+"__bubble";t.has(r)||(Vr(n,e,2,!1),t.add(r))}function $r(e,n,t){var r=0;n&&(r|=4),Vr(t,e,r,n)}var Ar="_reactListening"+Math.random().toString(36).slice(2);function Hr(e){if(!e[Ar]){e[Ar]=!0,o.forEach((function(n){"selectionchange"!==n&&(Fr.has(n)||$r(n,!1,e),$r(n,!0,e))}));var n=9===e.nodeType?e:e.ownerDocument;null===n||n[Ar]||(n[Ar]=!0,$r("selectionchange",!1,n))}}function Vr(e,n,t,r){switch(Gn(n)){case 1:var l=Qn;break;case 4:l=qn;break;default:l=Kn}t=l.bind(null,n,t,e),l=void 0,!Oe||"touchstart"!==n&&"touchmove"!==n&&"wheel"!==n||(l=!0),r?void 0!==l?e.addEventListener(n,t,{capture:!0,passive:l}):e.addEventListener(n,t,!0):void 0!==l?e.addEventListener(n,t,{passive:l}):e.addEventListener(n,t,!1)}function Br(e,n,t,r,l){var a=r;if(0===(1&n)&&0===(2&n)&&null!==r)e:for(;;){if(null===r)return;var o=r.tag;if(3===o||4===o){var u=r.stateNode.containerInfo;if(u===l||8===u.nodeType&&u.parentNode===l)break;if(4===o)for(o=r.return;null!==o;){var i=o.tag;if((3===i||4===i)&&((i=o.stateNode.containerInfo)===l||8===i.nodeType&&i.parentNode===l))return;o=o.return}for(;null!==u;){if(null===(o=yl(u)))return;if(5===(i=o.tag)||6===i){r=a=o;continue e}u=u.parentNode}}r=r.return}Te((function(){var r=a,l=we(t),o=[];e:{var u=Lr.get(e);if(void 0!==u){var i=ct,s=e;switch(e){case"keypress":if(0===tt(t))break e;case"keydown":case"keyup":i=Nt;break;case"focusin":s="focus",i=vt;break;case"focusout":s="blur",i=vt;break;case"beforeblur":case"afterblur":i=vt;break;case"click":if(2===t.button)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":i=mt;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":i=ht;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":i=zt;break;case _r:case Nr:case Pr:i=gt;break;case zr:i=Lt;break;case"scroll":i=dt;break;case"wheel":i=Rt;break;case"copy":case"cut":case"paste":i=bt;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":i=Pt}var c=0!==(4&n),f=!c&&"scroll"===e,d=c?null!==u?u+"Capture":null:u;c=[];for(var p,m=r;null!==m;){var h=(p=m).stateNode;if(5===p.tag&&null!==h&&(p=h,null!==d&&(null!=(h=Re(m,d))&&c.push(Wr(m,h,p)))),f)break;m=m.return}0<c.length&&(u=new i(u,s,null,t,l),o.push({event:u,listeners:c}))}}if(0===(7&n)){if(i="mouseout"===e||"pointerout"===e,(!(u="mouseover"===e||"pointerover"===e)||t===ke||!(s=t.relatedTarget||t.fromElement)||!yl(s)&&!s[ml])&&(i||u)&&(u=l.window===l?l:(u=l.ownerDocument)?u.defaultView||u.parentWindow:window,i?(i=r,null!==(s=(s=t.relatedTarget||t.toElement)?yl(s):null)&&(s!==(f=He(s))||5!==s.tag&&6!==s.tag)&&(s=null)):(i=null,s=r),i!==s)){if(c=mt,h="onMouseLeave",d="onMouseEnter",m="mouse","pointerout"!==e&&"pointerover"!==e||(c=Pt,h="onPointerLeave",d="onPointerEnter",m="pointer"),f=null==i?u:kl(i),p=null==s?u:kl(s),(u=new c(h,m+"leave",i,t,l)).target=f,u.relatedTarget=p,h=null,yl(l)===r&&((c=new c(d,m+"enter",s,t,l)).target=p,c.relatedTarget=f,h=c),f=h,i&&s)e:{for(d=s,m=0,p=c=i;p;p=qr(p))m++;for(p=0,h=d;h;h=qr(h))p++;for(;0<m-p;)c=qr(c),m--;for(;0<p-m;)d=qr(d),p--;for(;m--;){if(c===d||null!==d&&c===d.alternate)break e;c=qr(c),d=qr(d)}c=null}else c=null;null!==i&&Kr(o,u,i,c,!1),null!==s&&null!==f&&Kr(o,f,s,c,!0)}if("select"===(i=(u=r?kl(r):window).nodeName&&u.nodeName.toLowerCase())||"input"===i&&"file"===u.type)var v=Xt;else if(Bt(u))if(Gt)v=or;else{v=lr;var g=rr}else(i=u.nodeName)&&"input"===i.toLowerCase()&&("checkbox"===u.type||"radio"===u.type)&&(v=ar);switch(v&&(v=v(e,r))?Wt(o,v,t,l):(g&&g(e,u,r),"focusout"===e&&(g=u._wrapperState)&&g.controlled&&"number"===u.type&&ee(u,"number",u.value)),g=r?kl(r):window,e){case"focusin":(Bt(g)||"true"===g.contentEditable)&&(vr=g,gr=r,yr=null);break;case"focusout":yr=gr=vr=null;break;case"mousedown":br=!0;break;case"contextmenu":case"mouseup":case"dragend":br=!1,kr(o,t,l);break;case"selectionchange":if(hr)break;case"keydown":case"keyup":kr(o,t,l)}var y;if(Mt)e:{switch(e){case"compositionstart":var b="onCompositionStart";break e;case"compositionend":b="onCompositionEnd";break e;case"compositionupdate":b="onCompositionUpdate";break e}b=void 0}else Ht?$t(e,t)&&(b="onCompositionEnd"):"keydown"===e&&229===t.keyCode&&(b="onCompositionStart");b&&(It&&"ko"!==t.locale&&(Ht||"onCompositionStart"!==b?"onCompositionEnd"===b&&Ht&&(y=nt()):(Zn="value"in(Jn=l)?Jn.value:Jn.textContent,Ht=!0)),0<(g=Qr(r,b)).length&&(b=new kt(b,e,null,t,l),o.push({event:b,listeners:g}),y?b.data=y:null!==(y=At(t))&&(b.data=y))),(y=Ft?function(e,n){switch(e){case"compositionend":return At(n);case"keypress":return 32!==n.which?null:(Ut=!0,Dt);case"textInput":return(e=n.data)===Dt&&Ut?null:e;default:return null}}(e,t):function(e,n){if(Ht)return"compositionend"===e||!Mt&&$t(e,n)?(e=nt(),et=Zn=Jn=null,Ht=!1,e):null;switch(e){case"paste":default:return null;case"keypress":if(!(n.ctrlKey||n.altKey||n.metaKey)||n.ctrlKey&&n.altKey){if(n.char&&1<n.char.length)return n.char;if(n.which)return String.fromCharCode(n.which)}return null;case"compositionend":return It&&"ko"!==n.locale?null:n.data}}(e,t))&&(0<(r=Qr(r,"onBeforeInput")).length&&(l=new kt("onBeforeInput","beforeinput",null,t,l),o.push({event:l,listeners:r}),l.data=y))}Dr(o,n)}))}function Wr(e,n,t){return{instance:e,listener:n,currentTarget:t}}function Qr(e,n){for(var t=n+"Capture",r=[];null!==e;){var l=e,a=l.stateNode;5===l.tag&&null!==a&&(l=a,null!=(a=Re(e,t))&&r.unshift(Wr(e,a,l)),null!=(a=Re(e,n))&&r.push(Wr(e,a,l))),e=e.return}return r}function qr(e){if(null===e)return null;do{e=e.return}while(e&&5!==e.tag);return e||null}function Kr(e,n,t,r,l){for(var a=n._reactName,o=[];null!==t&&t!==r;){var u=t,i=u.alternate,s=u.stateNode;if(null!==i&&i===r)break;5===u.tag&&null!==s&&(u=s,l?null!=(i=Re(t,a))&&o.unshift(Wr(t,i,u)):l||null!=(i=Re(t,a))&&o.push(Wr(t,i,u))),t=t.return}0!==o.length&&e.push({event:n,listeners:o})}var Yr=/\r\n?/g,Xr=/\u0000|\uFFFD/g;function Gr(e){return("string"===typeof e?e:""+e).replace(Yr,"\n").replace(Xr,"")}function Jr(e,n,t){if(n=Gr(n),Gr(e)!==n&&t)throw Error(a(425))}function Zr(){}var el=null,nl=null;function tl(e,n){return"textarea"===e||"noscript"===e||"string"===typeof n.children||"number"===typeof n.children||"object"===typeof n.dangerouslySetInnerHTML&&null!==n.dangerouslySetInnerHTML&&null!=n.dangerouslySetInnerHTML.__html}var rl="function"===typeof setTimeout?setTimeout:void 0,ll="function"===typeof clearTimeout?clearTimeout:void 0,al="function"===typeof Promise?Promise:void 0,ol="function"===typeof queueMicrotask?queueMicrotask:"undefined"!==typeof al?function(e){return al.resolve(null).then(e).catch(ul)}:rl;function ul(e){setTimeout((function(){throw e}))}function il(e,n){var t=n,r=0;do{var l=t.nextSibling;if(e.removeChild(t),l&&8===l.nodeType)if("/$"===(t=l.data)){if(0===r)return e.removeChild(l),void Vn(n);r--}else"$"!==t&&"$?"!==t&&"$!"!==t||r++;t=l}while(t);Vn(n)}function sl(e){for(;null!=e;e=e.nextSibling){var n=e.nodeType;if(1===n||3===n)break;if(8===n){if("$"===(n=e.data)||"$!"===n||"$?"===n)break;if("/$"===n)return null}}return e}function cl(e){e=e.previousSibling;for(var n=0;e;){if(8===e.nodeType){var t=e.data;if("$"===t||"$!"===t||"$?"===t){if(0===n)return e;n--}else"/$"===t&&n++}e=e.previousSibling}return null}var fl=Math.random().toString(36).slice(2),dl="__reactFiber$"+fl,pl="__reactProps$"+fl,ml="__reactContainer$"+fl,hl="__reactEvents$"+fl,vl="__reactListeners$"+fl,gl="__reactHandles$"+fl;function yl(e){var n=e[dl];if(n)return n;for(var t=e.parentNode;t;){if(n=t[ml]||t[dl]){if(t=n.alternate,null!==n.child||null!==t&&null!==t.child)for(e=cl(e);null!==e;){if(t=e[dl])return t;e=cl(e)}return n}t=(e=t).parentNode}return null}function bl(e){return!(e=e[dl]||e[ml])||5!==e.tag&&6!==e.tag&&13!==e.tag&&3!==e.tag?null:e}function kl(e){if(5===e.tag||6===e.tag)return e.stateNode;throw Error(a(33))}function wl(e){return e[pl]||null}var Sl=[],xl=-1;function El(e){return{current:e}}function Cl(e){0>xl||(e.current=Sl[xl],Sl[xl]=null,xl--)}function _l(e,n){xl++,Sl[xl]=e.current,e.current=n}var Nl={},Pl=El(Nl),zl=El(!1),Ll=Nl;function Tl(e,n){var t=e.type.contextTypes;if(!t)return Nl;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===n)return r.__reactInternalMemoizedMaskedChildContext;var l,a={};for(l in t)a[l]=n[l];return r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=n,e.__reactInternalMemoizedMaskedChildContext=a),a}function Rl(e){return null!==(e=e.childContextTypes)&&void 0!==e}function Ol(){Cl(zl),Cl(Pl)}function Ml(e,n,t){if(Pl.current!==Nl)throw Error(a(168));_l(Pl,n),_l(zl,t)}function jl(e,n,t){var r=e.stateNode;if(n=n.childContextTypes,"function"!==typeof r.getChildContext)return t;for(var l in r=r.getChildContext())if(!(l in n))throw Error(a(108,V(e)||"Unknown",l));return I({},t,r)}function Fl(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||Nl,Ll=Pl.current,_l(Pl,e),_l(zl,zl.current),!0}function Il(e,n,t){var r=e.stateNode;if(!r)throw Error(a(169));t?(e=jl(e,n,Ll),r.__reactInternalMemoizedMergedChildContext=e,Cl(zl),Cl(Pl),_l(Pl,e)):Cl(zl),_l(zl,t)}var Dl=null,Ul=!1,$l=!1;function Al(e){null===Dl?Dl=[e]:Dl.push(e)}function Hl(){if(!$l&&null!==Dl){$l=!0;var e=0,n=kn;try{var t=Dl;for(kn=1;e<t.length;e++){var r=t[e];do{r=r(!0)}while(null!==r)}Dl=null,Ul=!1}catch(l){throw null!==Dl&&(Dl=Dl.slice(e+1)),qe(Ze,Hl),l}finally{kn=n,$l=!1}}return null}var Vl=[],Bl=0,Wl=null,Ql=0,ql=[],Kl=0,Yl=null,Xl=1,Gl="";function Jl(e,n){Vl[Bl++]=Ql,Vl[Bl++]=Wl,Wl=e,Ql=n}function Zl(e,n,t){ql[Kl++]=Xl,ql[Kl++]=Gl,ql[Kl++]=Yl,Yl=e;var r=Xl;e=Gl;var l=32-on(r)-1;r&=~(1<<l),t+=1;var a=32-on(n)+l;if(30<a){var o=l-l%5;a=(r&(1<<o)-1).toString(32),r>>=o,l-=o,Xl=1<<32-on(n)+l|t<<l|r,Gl=a+e}else Xl=1<<a|t<<l|r,Gl=e}function ea(e){null!==e.return&&(Jl(e,1),Zl(e,1,0))}function na(e){for(;e===Wl;)Wl=Vl[--Bl],Vl[Bl]=null,Ql=Vl[--Bl],Vl[Bl]=null;for(;e===Yl;)Yl=ql[--Kl],ql[Kl]=null,Gl=ql[--Kl],ql[Kl]=null,Xl=ql[--Kl],ql[Kl]=null}var ta=null,ra=null,la=!1,aa=null;function oa(e,n){var t=Ts(5,null,null,0);t.elementType="DELETED",t.stateNode=n,t.return=e,null===(n=e.deletions)?(e.deletions=[t],e.flags|=16):n.push(t)}function ua(e,n){switch(e.tag){case 5:var t=e.type;return null!==(n=1!==n.nodeType||t.toLowerCase()!==n.nodeName.toLowerCase()?null:n)&&(e.stateNode=n,ta=e,ra=sl(n.firstChild),!0);case 6:return null!==(n=""===e.pendingProps||3!==n.nodeType?null:n)&&(e.stateNode=n,ta=e,ra=null,!0);case 13:return null!==(n=8!==n.nodeType?null:n)&&(t=null!==Yl?{id:Xl,overflow:Gl}:null,e.memoizedState={dehydrated:n,treeContext:t,retryLane:1073741824},(t=Ts(18,null,null,0)).stateNode=n,t.return=e,e.child=t,ta=e,ra=null,!0);default:return!1}}function ia(e){return 0!==(1&e.mode)&&0===(128&e.flags)}function sa(e){if(la){var n=ra;if(n){var t=n;if(!ua(e,n)){if(ia(e))throw Error(a(418));n=sl(t.nextSibling);var r=ta;n&&ua(e,n)?oa(r,t):(e.flags=-4097&e.flags|2,la=!1,ta=e)}}else{if(ia(e))throw Error(a(418));e.flags=-4097&e.flags|2,la=!1,ta=e}}}function ca(e){for(e=e.return;null!==e&&5!==e.tag&&3!==e.tag&&13!==e.tag;)e=e.return;ta=e}function fa(e){if(e!==ta)return!1;if(!la)return ca(e),la=!0,!1;var n;if((n=3!==e.tag)&&!(n=5!==e.tag)&&(n="head"!==(n=e.type)&&"body"!==n&&!tl(e.type,e.memoizedProps)),n&&(n=ra)){if(ia(e))throw da(),Error(a(418));for(;n;)oa(e,n),n=sl(n.nextSibling)}if(ca(e),13===e.tag){if(!(e=null!==(e=e.memoizedState)?e.dehydrated:null))throw Error(a(317));e:{for(e=e.nextSibling,n=0;e;){if(8===e.nodeType){var t=e.data;if("/$"===t){if(0===n){ra=sl(e.nextSibling);break e}n--}else"$"!==t&&"$!"!==t&&"$?"!==t||n++}e=e.nextSibling}ra=null}}else ra=ta?sl(e.stateNode.nextSibling):null;return!0}function da(){for(var e=ra;e;)e=sl(e.nextSibling)}function pa(){ra=ta=null,la=!1}function ma(e){null===aa?aa=[e]:aa.push(e)}var ha=k.ReactCurrentBatchConfig;function va(e,n,t){if(null!==(e=t.ref)&&"function"!==typeof e&&"object"!==typeof e){if(t._owner){if(t=t._owner){if(1!==t.tag)throw Error(a(309));var r=t.stateNode}if(!r)throw Error(a(147,e));var l=r,o=""+e;return null!==n&&null!==n.ref&&"function"===typeof n.ref&&n.ref._stringRef===o?n.ref:(n=function(e){var n=l.refs;null===e?delete n[o]:n[o]=e},n._stringRef=o,n)}if("string"!==typeof e)throw Error(a(284));if(!t._owner)throw Error(a(290,e))}return e}function ga(e,n){throw e=Object.prototype.toString.call(n),Error(a(31,"[object Object]"===e?"object with keys {"+Object.keys(n).join(", ")+"}":e))}function ya(e){return(0,e._init)(e._payload)}function ba(e){function n(n,t){if(e){var r=n.deletions;null===r?(n.deletions=[t],n.flags|=16):r.push(t)}}function t(t,r){if(!e)return null;for(;null!==r;)n(t,r),r=r.sibling;return null}function r(e,n){for(e=new Map;null!==n;)null!==n.key?e.set(n.key,n):e.set(n.index,n),n=n.sibling;return e}function l(e,n){return(e=Os(e,n)).index=0,e.sibling=null,e}function o(n,t,r){return n.index=r,e?null!==(r=n.alternate)?(r=r.index)<t?(n.flags|=2,t):r:(n.flags|=2,t):(n.flags|=1048576,t)}function u(n){return e&&null===n.alternate&&(n.flags|=2),n}function i(e,n,t,r){return null===n||6!==n.tag?((n=Is(t,e.mode,r)).return=e,n):((n=l(n,t)).return=e,n)}function s(e,n,t,r){var a=t.type;return a===x?f(e,n,t.props.children,r,t.key):null!==n&&(n.elementType===a||"object"===typeof a&&null!==a&&a.$$typeof===R&&ya(a)===n.type)?((r=l(n,t.props)).ref=va(e,n,t),r.return=e,r):((r=Ms(t.type,t.key,t.props,null,e.mode,r)).ref=va(e,n,t),r.return=e,r)}function c(e,n,t,r){return null===n||4!==n.tag||n.stateNode.containerInfo!==t.containerInfo||n.stateNode.implementation!==t.implementation?((n=Ds(t,e.mode,r)).return=e,n):((n=l(n,t.children||[])).return=e,n)}function f(e,n,t,r,a){return null===n||7!==n.tag?((n=js(t,e.mode,r,a)).return=e,n):((n=l(n,t)).return=e,n)}function d(e,n,t){if("string"===typeof n&&""!==n||"number"===typeof n)return(n=Is(""+n,e.mode,t)).return=e,n;if("object"===typeof n&&null!==n){switch(n.$$typeof){case w:return(t=Ms(n.type,n.key,n.props,null,e.mode,t)).ref=va(e,null,n),t.return=e,t;case S:return(n=Ds(n,e.mode,t)).return=e,n;case R:return d(e,(0,n._init)(n._payload),t)}if(ne(n)||j(n))return(n=js(n,e.mode,t,null)).return=e,n;ga(e,n)}return null}function p(e,n,t,r){var l=null!==n?n.key:null;if("string"===typeof t&&""!==t||"number"===typeof t)return null!==l?null:i(e,n,""+t,r);if("object"===typeof t&&null!==t){switch(t.$$typeof){case w:return t.key===l?s(e,n,t,r):null;case S:return t.key===l?c(e,n,t,r):null;case R:return p(e,n,(l=t._init)(t._payload),r)}if(ne(t)||j(t))return null!==l?null:f(e,n,t,r,null);ga(e,t)}return null}function m(e,n,t,r,l){if("string"===typeof r&&""!==r||"number"===typeof r)return i(n,e=e.get(t)||null,""+r,l);if("object"===typeof r&&null!==r){switch(r.$$typeof){case w:return s(n,e=e.get(null===r.key?t:r.key)||null,r,l);case S:return c(n,e=e.get(null===r.key?t:r.key)||null,r,l);case R:return m(e,n,t,(0,r._init)(r._payload),l)}if(ne(r)||j(r))return f(n,e=e.get(t)||null,r,l,null);ga(n,r)}return null}function h(l,a,u,i){for(var s=null,c=null,f=a,h=a=0,v=null;null!==f&&h<u.length;h++){f.index>h?(v=f,f=null):v=f.sibling;var g=p(l,f,u[h],i);if(null===g){null===f&&(f=v);break}e&&f&&null===g.alternate&&n(l,f),a=o(g,a,h),null===c?s=g:c.sibling=g,c=g,f=v}if(h===u.length)return t(l,f),la&&Jl(l,h),s;if(null===f){for(;h<u.length;h++)null!==(f=d(l,u[h],i))&&(a=o(f,a,h),null===c?s=f:c.sibling=f,c=f);return la&&Jl(l,h),s}for(f=r(l,f);h<u.length;h++)null!==(v=m(f,l,h,u[h],i))&&(e&&null!==v.alternate&&f.delete(null===v.key?h:v.key),a=o(v,a,h),null===c?s=v:c.sibling=v,c=v);return e&&f.forEach((function(e){return n(l,e)})),la&&Jl(l,h),s}function v(l,u,i,s){var c=j(i);if("function"!==typeof c)throw Error(a(150));if(null==(i=c.call(i)))throw Error(a(151));for(var f=c=null,h=u,v=u=0,g=null,y=i.next();null!==h&&!y.done;v++,y=i.next()){h.index>v?(g=h,h=null):g=h.sibling;var b=p(l,h,y.value,s);if(null===b){null===h&&(h=g);break}e&&h&&null===b.alternate&&n(l,h),u=o(b,u,v),null===f?c=b:f.sibling=b,f=b,h=g}if(y.done)return t(l,h),la&&Jl(l,v),c;if(null===h){for(;!y.done;v++,y=i.next())null!==(y=d(l,y.value,s))&&(u=o(y,u,v),null===f?c=y:f.sibling=y,f=y);return la&&Jl(l,v),c}for(h=r(l,h);!y.done;v++,y=i.next())null!==(y=m(h,l,v,y.value,s))&&(e&&null!==y.alternate&&h.delete(null===y.key?v:y.key),u=o(y,u,v),null===f?c=y:f.sibling=y,f=y);return e&&h.forEach((function(e){return n(l,e)})),la&&Jl(l,v),c}return function e(r,a,o,i){if("object"===typeof o&&null!==o&&o.type===x&&null===o.key&&(o=o.props.children),"object"===typeof o&&null!==o){switch(o.$$typeof){case w:e:{for(var s=o.key,c=a;null!==c;){if(c.key===s){if((s=o.type)===x){if(7===c.tag){t(r,c.sibling),(a=l(c,o.props.children)).return=r,r=a;break e}}else if(c.elementType===s||"object"===typeof s&&null!==s&&s.$$typeof===R&&ya(s)===c.type){t(r,c.sibling),(a=l(c,o.props)).ref=va(r,c,o),a.return=r,r=a;break e}t(r,c);break}n(r,c),c=c.sibling}o.type===x?((a=js(o.props.children,r.mode,i,o.key)).return=r,r=a):((i=Ms(o.type,o.key,o.props,null,r.mode,i)).ref=va(r,a,o),i.return=r,r=i)}return u(r);case S:e:{for(c=o.key;null!==a;){if(a.key===c){if(4===a.tag&&a.stateNode.containerInfo===o.containerInfo&&a.stateNode.implementation===o.implementation){t(r,a.sibling),(a=l(a,o.children||[])).return=r,r=a;break e}t(r,a);break}n(r,a),a=a.sibling}(a=Ds(o,r.mode,i)).return=r,r=a}return u(r);case R:return e(r,a,(c=o._init)(o._payload),i)}if(ne(o))return h(r,a,o,i);if(j(o))return v(r,a,o,i);ga(r,o)}return"string"===typeof o&&""!==o||"number"===typeof o?(o=""+o,null!==a&&6===a.tag?(t(r,a.sibling),(a=l(a,o)).return=r,r=a):(t(r,a),(a=Is(o,r.mode,i)).return=r,r=a),u(r)):t(r,a)}}var ka=ba(!0),wa=ba(!1),Sa=El(null),xa=null,Ea=null,Ca=null;function _a(){Ca=Ea=xa=null}function Na(e){var n=Sa.current;Cl(Sa),e._currentValue=n}function Pa(e,n,t){for(;null!==e;){var r=e.alternate;if((e.childLanes&n)!==n?(e.childLanes|=n,null!==r&&(r.childLanes|=n)):null!==r&&(r.childLanes&n)!==n&&(r.childLanes|=n),e===t)break;e=e.return}}function za(e,n){xa=e,Ca=Ea=null,null!==(e=e.dependencies)&&null!==e.firstContext&&(0!==(e.lanes&n)&&(bu=!0),e.firstContext=null)}function La(e){var n=e._currentValue;if(Ca!==e)if(e={context:e,memoizedValue:n,next:null},null===Ea){if(null===xa)throw Error(a(308));Ea=e,xa.dependencies={lanes:0,firstContext:e}}else Ea=Ea.next=e;return n}var Ta=null;function Ra(e){null===Ta?Ta=[e]:Ta.push(e)}function Oa(e,n,t,r){var l=n.interleaved;return null===l?(t.next=t,Ra(n)):(t.next=l.next,l.next=t),n.interleaved=t,Ma(e,r)}function Ma(e,n){e.lanes|=n;var t=e.alternate;for(null!==t&&(t.lanes|=n),t=e,e=e.return;null!==e;)e.childLanes|=n,null!==(t=e.alternate)&&(t.childLanes|=n),t=e,e=e.return;return 3===t.tag?t.stateNode:null}var ja=!1;function Fa(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Ia(e,n){e=e.updateQueue,n.updateQueue===e&&(n.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function Da(e,n){return{eventTime:e,lane:n,tag:0,payload:null,callback:null,next:null}}function Ua(e,n,t){var r=e.updateQueue;if(null===r)return null;if(r=r.shared,0!==(2&Pi)){var l=r.pending;return null===l?n.next=n:(n.next=l.next,l.next=n),r.pending=n,Ma(e,t)}return null===(l=r.interleaved)?(n.next=n,Ra(r)):(n.next=l.next,l.next=n),r.interleaved=n,Ma(e,t)}function $a(e,n,t){if(null!==(n=n.updateQueue)&&(n=n.shared,0!==(4194240&t))){var r=n.lanes;t|=r&=e.pendingLanes,n.lanes=t,bn(e,t)}}function Aa(e,n){var t=e.updateQueue,r=e.alternate;if(null!==r&&t===(r=r.updateQueue)){var l=null,a=null;if(null!==(t=t.firstBaseUpdate)){do{var o={eventTime:t.eventTime,lane:t.lane,tag:t.tag,payload:t.payload,callback:t.callback,next:null};null===a?l=a=o:a=a.next=o,t=t.next}while(null!==t);null===a?l=a=n:a=a.next=n}else l=a=n;return t={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:a,shared:r.shared,effects:r.effects},void(e.updateQueue=t)}null===(e=t.lastBaseUpdate)?t.firstBaseUpdate=n:e.next=n,t.lastBaseUpdate=n}function Ha(e,n,t,r){var l=e.updateQueue;ja=!1;var a=l.firstBaseUpdate,o=l.lastBaseUpdate,u=l.shared.pending;if(null!==u){l.shared.pending=null;var i=u,s=i.next;i.next=null,null===o?a=s:o.next=s,o=i;var c=e.alternate;null!==c&&((u=(c=c.updateQueue).lastBaseUpdate)!==o&&(null===u?c.firstBaseUpdate=s:u.next=s,c.lastBaseUpdate=i))}if(null!==a){var f=l.baseState;for(o=0,c=s=i=null,u=a;;){var d=u.lane,p=u.eventTime;if((r&d)===d){null!==c&&(c=c.next={eventTime:p,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var m=e,h=u;switch(d=n,p=t,h.tag){case 1:if("function"===typeof(m=h.payload)){f=m.call(p,f,d);break e}f=m;break e;case 3:m.flags=-65537&m.flags|128;case 0:if(null===(d="function"===typeof(m=h.payload)?m.call(p,f,d):m)||void 0===d)break e;f=I({},f,d);break e;case 2:ja=!0}}null!==u.callback&&0!==u.lane&&(e.flags|=64,null===(d=l.effects)?l.effects=[u]:d.push(u))}else p={eventTime:p,lane:d,tag:u.tag,payload:u.payload,callback:u.callback,next:null},null===c?(s=c=p,i=f):c=c.next=p,o|=d;if(null===(u=u.next)){if(null===(u=l.shared.pending))break;u=(d=u).next,d.next=null,l.lastBaseUpdate=d,l.shared.pending=null}}if(null===c&&(i=f),l.baseState=i,l.firstBaseUpdate=s,l.lastBaseUpdate=c,null!==(n=l.shared.interleaved)){l=n;do{o|=l.lane,l=l.next}while(l!==n)}else null===a&&(l.shared.lanes=0);Fi|=o,e.lanes=o,e.memoizedState=f}}function Va(e,n,t){if(e=n.effects,n.effects=null,null!==e)for(n=0;n<e.length;n++){var r=e[n],l=r.callback;if(null!==l){if(r.callback=null,r=t,"function"!==typeof l)throw Error(a(191,l));l.call(r)}}}var Ba={},Wa=El(Ba),Qa=El(Ba),qa=El(Ba);function Ka(e){if(e===Ba)throw Error(a(174));return e}function Ya(e,n){switch(_l(qa,n),_l(Qa,e),_l(Wa,Ba),e=n.nodeType){case 9:case 11:n=(n=n.documentElement)?n.namespaceURI:ie(null,"");break;default:n=ie(n=(e=8===e?n.parentNode:n).namespaceURI||null,e=e.tagName)}Cl(Wa),_l(Wa,n)}function Xa(){Cl(Wa),Cl(Qa),Cl(qa)}function Ga(e){Ka(qa.current);var n=Ka(Wa.current),t=ie(n,e.type);n!==t&&(_l(Qa,e),_l(Wa,t))}function Ja(e){Qa.current===e&&(Cl(Wa),Cl(Qa))}var Za=El(0);function eo(e){for(var n=e;null!==n;){if(13===n.tag){var t=n.memoizedState;if(null!==t&&(null===(t=t.dehydrated)||"$?"===t.data||"$!"===t.data))return n}else if(19===n.tag&&void 0!==n.memoizedProps.revealOrder){if(0!==(128&n.flags))return n}else if(null!==n.child){n.child.return=n,n=n.child;continue}if(n===e)break;for(;null===n.sibling;){if(null===n.return||n.return===e)return null;n=n.return}n.sibling.return=n.return,n=n.sibling}return null}var no=[];function to(){for(var e=0;e<no.length;e++)no[e]._workInProgressVersionPrimary=null;no.length=0}var ro=k.ReactCurrentDispatcher,lo=k.ReactCurrentBatchConfig,ao=0,oo=null,uo=null,io=null,so=!1,co=!1,fo=0,po=0;function mo(){throw Error(a(321))}function ho(e,n){if(null===n)return!1;for(var t=0;t<n.length&&t<e.length;t++)if(!ur(e[t],n[t]))return!1;return!0}function vo(e,n,t,r,l,o){if(ao=o,oo=n,n.memoizedState=null,n.updateQueue=null,n.lanes=0,ro.current=null===e||null===e.memoizedState?Zo:eu,e=t(r,l),co){o=0;do{if(co=!1,fo=0,25<=o)throw Error(a(301));o+=1,io=uo=null,n.updateQueue=null,ro.current=nu,e=t(r,l)}while(co)}if(ro.current=Jo,n=null!==uo&&null!==uo.next,ao=0,io=uo=oo=null,so=!1,n)throw Error(a(300));return e}function go(){var e=0!==fo;return fo=0,e}function yo(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return null===io?oo.memoizedState=io=e:io=io.next=e,io}function bo(){if(null===uo){var e=oo.alternate;e=null!==e?e.memoizedState:null}else e=uo.next;var n=null===io?oo.memoizedState:io.next;if(null!==n)io=n,uo=e;else{if(null===e)throw Error(a(310));e={memoizedState:(uo=e).memoizedState,baseState:uo.baseState,baseQueue:uo.baseQueue,queue:uo.queue,next:null},null===io?oo.memoizedState=io=e:io=io.next=e}return io}function ko(e,n){return"function"===typeof n?n(e):n}function wo(e){var n=bo(),t=n.queue;if(null===t)throw Error(a(311));t.lastRenderedReducer=e;var r=uo,l=r.baseQueue,o=t.pending;if(null!==o){if(null!==l){var u=l.next;l.next=o.next,o.next=u}r.baseQueue=l=o,t.pending=null}if(null!==l){o=l.next,r=r.baseState;var i=u=null,s=null,c=o;do{var f=c.lane;if((ao&f)===f)null!==s&&(s=s.next={lane:0,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null}),r=c.hasEagerState?c.eagerState:e(r,c.action);else{var d={lane:f,action:c.action,hasEagerState:c.hasEagerState,eagerState:c.eagerState,next:null};null===s?(i=s=d,u=r):s=s.next=d,oo.lanes|=f,Fi|=f}c=c.next}while(null!==c&&c!==o);null===s?u=r:s.next=i,ur(r,n.memoizedState)||(bu=!0),n.memoizedState=r,n.baseState=u,n.baseQueue=s,t.lastRenderedState=r}if(null!==(e=t.interleaved)){l=e;do{o=l.lane,oo.lanes|=o,Fi|=o,l=l.next}while(l!==e)}else null===l&&(t.lanes=0);return[n.memoizedState,t.dispatch]}function So(e){var n=bo(),t=n.queue;if(null===t)throw Error(a(311));t.lastRenderedReducer=e;var r=t.dispatch,l=t.pending,o=n.memoizedState;if(null!==l){t.pending=null;var u=l=l.next;do{o=e(o,u.action),u=u.next}while(u!==l);ur(o,n.memoizedState)||(bu=!0),n.memoizedState=o,null===n.baseQueue&&(n.baseState=o),t.lastRenderedState=o}return[o,r]}function xo(){}function Eo(e,n){var t=oo,r=bo(),l=n(),o=!ur(r.memoizedState,l);if(o&&(r.memoizedState=l,bu=!0),r=r.queue,Fo(No.bind(null,t,r,e),[e]),r.getSnapshot!==n||o||null!==io&&1&io.memoizedState.tag){if(t.flags|=2048,To(9,_o.bind(null,t,r,l,n),void 0,null),null===zi)throw Error(a(349));0!==(30&ao)||Co(t,n,l)}return l}function Co(e,n,t){e.flags|=16384,e={getSnapshot:n,value:t},null===(n=oo.updateQueue)?(n={lastEffect:null,stores:null},oo.updateQueue=n,n.stores=[e]):null===(t=n.stores)?n.stores=[e]:t.push(e)}function _o(e,n,t,r){n.value=t,n.getSnapshot=r,Po(n)&&zo(e)}function No(e,n,t){return t((function(){Po(n)&&zo(e)}))}function Po(e){var n=e.getSnapshot;e=e.value;try{var t=n();return!ur(e,t)}catch(r){return!0}}function zo(e){var n=Ma(e,1);null!==n&&ts(n,e,1,-1)}function Lo(e){var n=yo();return"function"===typeof e&&(e=e()),n.memoizedState=n.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:ko,lastRenderedState:e},n.queue=e,e=e.dispatch=Ko.bind(null,oo,e),[n.memoizedState,e]}function To(e,n,t,r){return e={tag:e,create:n,destroy:t,deps:r,next:null},null===(n=oo.updateQueue)?(n={lastEffect:null,stores:null},oo.updateQueue=n,n.lastEffect=e.next=e):null===(t=n.lastEffect)?n.lastEffect=e.next=e:(r=t.next,t.next=e,e.next=r,n.lastEffect=e),e}function Ro(){return bo().memoizedState}function Oo(e,n,t,r){var l=yo();oo.flags|=e,l.memoizedState=To(1|n,t,void 0,void 0===r?null:r)}function Mo(e,n,t,r){var l=bo();r=void 0===r?null:r;var a=void 0;if(null!==uo){var o=uo.memoizedState;if(a=o.destroy,null!==r&&ho(r,o.deps))return void(l.memoizedState=To(n,t,a,r))}oo.flags|=e,l.memoizedState=To(1|n,t,a,r)}function jo(e,n){return Oo(8390656,8,e,n)}function Fo(e,n){return Mo(2048,8,e,n)}function Io(e,n){return Mo(4,2,e,n)}function Do(e,n){return Mo(4,4,e,n)}function Uo(e,n){return"function"===typeof n?(e=e(),n(e),function(){n(null)}):null!==n&&void 0!==n?(e=e(),n.current=e,function(){n.current=null}):void 0}function $o(e,n,t){return t=null!==t&&void 0!==t?t.concat([e]):null,Mo(4,4,Uo.bind(null,n,e),t)}function Ao(){}function Ho(e,n){var t=bo();n=void 0===n?null:n;var r=t.memoizedState;return null!==r&&null!==n&&ho(n,r[1])?r[0]:(t.memoizedState=[e,n],e)}function Vo(e,n){var t=bo();n=void 0===n?null:n;var r=t.memoizedState;return null!==r&&null!==n&&ho(n,r[1])?r[0]:(e=e(),t.memoizedState=[e,n],e)}function Bo(e,n,t){return 0===(21&ao)?(e.baseState&&(e.baseState=!1,bu=!0),e.memoizedState=t):(ur(t,n)||(t=vn(),oo.lanes|=t,Fi|=t,e.baseState=!0),n)}function Wo(e,n){var t=kn;kn=0!==t&&4>t?t:4,e(!0);var r=lo.transition;lo.transition={};try{e(!1),n()}finally{kn=t,lo.transition=r}}function Qo(){return bo().memoizedState}function qo(e,n,t){var r=ns(e);if(t={lane:r,action:t,hasEagerState:!1,eagerState:null,next:null},Yo(e))Xo(n,t);else if(null!==(t=Oa(e,n,t,r))){ts(t,e,r,es()),Go(t,n,r)}}function Ko(e,n,t){var r=ns(e),l={lane:r,action:t,hasEagerState:!1,eagerState:null,next:null};if(Yo(e))Xo(n,l);else{var a=e.alternate;if(0===e.lanes&&(null===a||0===a.lanes)&&null!==(a=n.lastRenderedReducer))try{var o=n.lastRenderedState,u=a(o,t);if(l.hasEagerState=!0,l.eagerState=u,ur(u,o)){var i=n.interleaved;return null===i?(l.next=l,Ra(n)):(l.next=i.next,i.next=l),void(n.interleaved=l)}}catch(s){}null!==(t=Oa(e,n,l,r))&&(ts(t,e,r,l=es()),Go(t,n,r))}}function Yo(e){var n=e.alternate;return e===oo||null!==n&&n===oo}function Xo(e,n){co=so=!0;var t=e.pending;null===t?n.next=n:(n.next=t.next,t.next=n),e.pending=n}function Go(e,n,t){if(0!==(4194240&t)){var r=n.lanes;t|=r&=e.pendingLanes,n.lanes=t,bn(e,t)}}var Jo={readContext:La,useCallback:mo,useContext:mo,useEffect:mo,useImperativeHandle:mo,useInsertionEffect:mo,useLayoutEffect:mo,useMemo:mo,useReducer:mo,useRef:mo,useState:mo,useDebugValue:mo,useDeferredValue:mo,useTransition:mo,useMutableSource:mo,useSyncExternalStore:mo,useId:mo,unstable_isNewReconciler:!1},Zo={readContext:La,useCallback:function(e,n){return yo().memoizedState=[e,void 0===n?null:n],e},useContext:La,useEffect:jo,useImperativeHandle:function(e,n,t){return t=null!==t&&void 0!==t?t.concat([e]):null,Oo(4194308,4,Uo.bind(null,n,e),t)},useLayoutEffect:function(e,n){return Oo(4194308,4,e,n)},useInsertionEffect:function(e,n){return Oo(4,2,e,n)},useMemo:function(e,n){var t=yo();return n=void 0===n?null:n,e=e(),t.memoizedState=[e,n],e},useReducer:function(e,n,t){var r=yo();return n=void 0!==t?t(n):n,r.memoizedState=r.baseState=n,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:n},r.queue=e,e=e.dispatch=qo.bind(null,oo,e),[r.memoizedState,e]},useRef:function(e){return e={current:e},yo().memoizedState=e},useState:Lo,useDebugValue:Ao,useDeferredValue:function(e){return yo().memoizedState=e},useTransition:function(){var e=Lo(!1),n=e[0];return e=Wo.bind(null,e[1]),yo().memoizedState=e,[n,e]},useMutableSource:function(){},useSyncExternalStore:function(e,n,t){var r=oo,l=yo();if(la){if(void 0===t)throw Error(a(407));t=t()}else{if(t=n(),null===zi)throw Error(a(349));0!==(30&ao)||Co(r,n,t)}l.memoizedState=t;var o={value:t,getSnapshot:n};return l.queue=o,jo(No.bind(null,r,o,e),[e]),r.flags|=2048,To(9,_o.bind(null,r,o,t,n),void 0,null),t},useId:function(){var e=yo(),n=zi.identifierPrefix;if(la){var t=Gl;n=":"+n+"R"+(t=(Xl&~(1<<32-on(Xl)-1)).toString(32)+t),0<(t=fo++)&&(n+="H"+t.toString(32)),n+=":"}else n=":"+n+"r"+(t=po++).toString(32)+":";return e.memoizedState=n},unstable_isNewReconciler:!1},eu={readContext:La,useCallback:Ho,useContext:La,useEffect:Fo,useImperativeHandle:$o,useInsertionEffect:Io,useLayoutEffect:Do,useMemo:Vo,useReducer:wo,useRef:Ro,useState:function(){return wo(ko)},useDebugValue:Ao,useDeferredValue:function(e){return Bo(bo(),uo.memoizedState,e)},useTransition:function(){return[wo(ko)[0],bo().memoizedState]},useMutableSource:xo,useSyncExternalStore:Eo,useId:Qo,unstable_isNewReconciler:!1},nu={readContext:La,useCallback:Ho,useContext:La,useEffect:Fo,useImperativeHandle:$o,useInsertionEffect:Io,useLayoutEffect:Do,useMemo:Vo,useReducer:So,useRef:Ro,useState:function(){return So(ko)},useDebugValue:Ao,useDeferredValue:function(e){var n=bo();return null===uo?n.memoizedState=e:Bo(n,uo.memoizedState,e)},useTransition:function(){return[So(ko)[0],bo().memoizedState]},useMutableSource:xo,useSyncExternalStore:Eo,useId:Qo,unstable_isNewReconciler:!1};function tu(e,n){if(e&&e.defaultProps){for(var t in n=I({},n),e=e.defaultProps)void 0===n[t]&&(n[t]=e[t]);return n}return n}function ru(e,n,t,r){t=null===(t=t(r,n=e.memoizedState))||void 0===t?n:I({},n,t),e.memoizedState=t,0===e.lanes&&(e.updateQueue.baseState=t)}var lu={isMounted:function(e){return!!(e=e._reactInternals)&&He(e)===e},enqueueSetState:function(e,n,t){e=e._reactInternals;var r=es(),l=ns(e),a=Da(r,l);a.payload=n,void 0!==t&&null!==t&&(a.callback=t),null!==(n=Ua(e,a,l))&&(ts(n,e,l,r),$a(n,e,l))},enqueueReplaceState:function(e,n,t){e=e._reactInternals;var r=es(),l=ns(e),a=Da(r,l);a.tag=1,a.payload=n,void 0!==t&&null!==t&&(a.callback=t),null!==(n=Ua(e,a,l))&&(ts(n,e,l,r),$a(n,e,l))},enqueueForceUpdate:function(e,n){e=e._reactInternals;var t=es(),r=ns(e),l=Da(t,r);l.tag=2,void 0!==n&&null!==n&&(l.callback=n),null!==(n=Ua(e,l,r))&&(ts(n,e,r,t),$a(n,e,r))}};function au(e,n,t,r,l,a,o){return"function"===typeof(e=e.stateNode).shouldComponentUpdate?e.shouldComponentUpdate(r,a,o):!n.prototype||!n.prototype.isPureReactComponent||(!ir(t,r)||!ir(l,a))}function ou(e,n,t){var r=!1,l=Nl,a=n.contextType;return"object"===typeof a&&null!==a?a=La(a):(l=Rl(n)?Ll:Pl.current,a=(r=null!==(r=n.contextTypes)&&void 0!==r)?Tl(e,l):Nl),n=new n(t,a),e.memoizedState=null!==n.state&&void 0!==n.state?n.state:null,n.updater=lu,e.stateNode=n,n._reactInternals=e,r&&((e=e.stateNode).__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=a),n}function uu(e,n,t,r){e=n.state,"function"===typeof n.componentWillReceiveProps&&n.componentWillReceiveProps(t,r),"function"===typeof n.UNSAFE_componentWillReceiveProps&&n.UNSAFE_componentWillReceiveProps(t,r),n.state!==e&&lu.enqueueReplaceState(n,n.state,null)}function iu(e,n,t,r){var l=e.stateNode;l.props=t,l.state=e.memoizedState,l.refs={},Fa(e);var a=n.contextType;"object"===typeof a&&null!==a?l.context=La(a):(a=Rl(n)?Ll:Pl.current,l.context=Tl(e,a)),l.state=e.memoizedState,"function"===typeof(a=n.getDerivedStateFromProps)&&(ru(e,n,a,t),l.state=e.memoizedState),"function"===typeof n.getDerivedStateFromProps||"function"===typeof l.getSnapshotBeforeUpdate||"function"!==typeof l.UNSAFE_componentWillMount&&"function"!==typeof l.componentWillMount||(n=l.state,"function"===typeof l.componentWillMount&&l.componentWillMount(),"function"===typeof l.UNSAFE_componentWillMount&&l.UNSAFE_componentWillMount(),n!==l.state&&lu.enqueueReplaceState(l,l.state,null),Ha(e,t,l,r),l.state=e.memoizedState),"function"===typeof l.componentDidMount&&(e.flags|=4194308)}function su(e,n){try{var t="",r=n;do{t+=A(r),r=r.return}while(r);var l=t}catch(a){l="\nError generating stack: "+a.message+"\n"+a.stack}return{value:e,source:n,stack:l,digest:null}}function cu(e,n,t){return{value:e,source:null,stack:null!=t?t:null,digest:null!=n?n:null}}function fu(e,n){try{console.error(n.value)}catch(t){setTimeout((function(){throw t}))}}var du="function"===typeof WeakMap?WeakMap:Map;function pu(e,n,t){(t=Da(-1,t)).tag=3,t.payload={element:null};var r=n.value;return t.callback=function(){Bi||(Bi=!0,Wi=r),fu(0,n)},t}function mu(e,n,t){(t=Da(-1,t)).tag=3;var r=e.type.getDerivedStateFromError;if("function"===typeof r){var l=n.value;t.payload=function(){return r(l)},t.callback=function(){fu(0,n)}}var a=e.stateNode;return null!==a&&"function"===typeof a.componentDidCatch&&(t.callback=function(){fu(0,n),"function"!==typeof r&&(null===Qi?Qi=new Set([this]):Qi.add(this));var e=n.stack;this.componentDidCatch(n.value,{componentStack:null!==e?e:""})}),t}function hu(e,n,t){var r=e.pingCache;if(null===r){r=e.pingCache=new du;var l=new Set;r.set(n,l)}else void 0===(l=r.get(n))&&(l=new Set,r.set(n,l));l.has(t)||(l.add(t),e=Cs.bind(null,e,n,t),n.then(e,e))}function vu(e){do{var n;if((n=13===e.tag)&&(n=null===(n=e.memoizedState)||null!==n.dehydrated),n)return e;e=e.return}while(null!==e);return null}function gu(e,n,t,r,l){return 0===(1&e.mode)?(e===n?e.flags|=65536:(e.flags|=128,t.flags|=131072,t.flags&=-52805,1===t.tag&&(null===t.alternate?t.tag=17:((n=Da(-1,1)).tag=2,Ua(t,n,1))),t.lanes|=1),e):(e.flags|=65536,e.lanes=l,e)}var yu=k.ReactCurrentOwner,bu=!1;function ku(e,n,t,r){n.child=null===e?wa(n,null,t,r):ka(n,e.child,t,r)}function wu(e,n,t,r,l){t=t.render;var a=n.ref;return za(n,l),r=vo(e,n,t,r,a,l),t=go(),null===e||bu?(la&&t&&ea(n),n.flags|=1,ku(e,n,r,l),n.child):(n.updateQueue=e.updateQueue,n.flags&=-2053,e.lanes&=~l,Bu(e,n,l))}function Su(e,n,t,r,l){if(null===e){var a=t.type;return"function"!==typeof a||Rs(a)||void 0!==a.defaultProps||null!==t.compare||void 0!==t.defaultProps?((e=Ms(t.type,null,r,n,n.mode,l)).ref=n.ref,e.return=n,n.child=e):(n.tag=15,n.type=a,xu(e,n,a,r,l))}if(a=e.child,0===(e.lanes&l)){var o=a.memoizedProps;if((t=null!==(t=t.compare)?t:ir)(o,r)&&e.ref===n.ref)return Bu(e,n,l)}return n.flags|=1,(e=Os(a,r)).ref=n.ref,e.return=n,n.child=e}function xu(e,n,t,r,l){if(null!==e){var a=e.memoizedProps;if(ir(a,r)&&e.ref===n.ref){if(bu=!1,n.pendingProps=r=a,0===(e.lanes&l))return n.lanes=e.lanes,Bu(e,n,l);0!==(131072&e.flags)&&(bu=!0)}}return _u(e,n,t,r,l)}function Eu(e,n,t){var r=n.pendingProps,l=r.children,a=null!==e?e.memoizedState:null;if("hidden"===r.mode)if(0===(1&n.mode))n.memoizedState={baseLanes:0,cachePool:null,transitions:null},_l(Oi,Ri),Ri|=t;else{if(0===(1073741824&t))return e=null!==a?a.baseLanes|t:t,n.lanes=n.childLanes=1073741824,n.memoizedState={baseLanes:e,cachePool:null,transitions:null},n.updateQueue=null,_l(Oi,Ri),Ri|=e,null;n.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=null!==a?a.baseLanes:t,_l(Oi,Ri),Ri|=r}else null!==a?(r=a.baseLanes|t,n.memoizedState=null):r=t,_l(Oi,Ri),Ri|=r;return ku(e,n,l,t),n.child}function Cu(e,n){var t=n.ref;(null===e&&null!==t||null!==e&&e.ref!==t)&&(n.flags|=512,n.flags|=2097152)}function _u(e,n,t,r,l){var a=Rl(t)?Ll:Pl.current;return a=Tl(n,a),za(n,l),t=vo(e,n,t,r,a,l),r=go(),null===e||bu?(la&&r&&ea(n),n.flags|=1,ku(e,n,t,l),n.child):(n.updateQueue=e.updateQueue,n.flags&=-2053,e.lanes&=~l,Bu(e,n,l))}function Nu(e,n,t,r,l){if(Rl(t)){var a=!0;Fl(n)}else a=!1;if(za(n,l),null===n.stateNode)Vu(e,n),ou(n,t,r),iu(n,t,r,l),r=!0;else if(null===e){var o=n.stateNode,u=n.memoizedProps;o.props=u;var i=o.context,s=t.contextType;"object"===typeof s&&null!==s?s=La(s):s=Tl(n,s=Rl(t)?Ll:Pl.current);var c=t.getDerivedStateFromProps,f="function"===typeof c||"function"===typeof o.getSnapshotBeforeUpdate;f||"function"!==typeof o.UNSAFE_componentWillReceiveProps&&"function"!==typeof o.componentWillReceiveProps||(u!==r||i!==s)&&uu(n,o,r,s),ja=!1;var d=n.memoizedState;o.state=d,Ha(n,r,o,l),i=n.memoizedState,u!==r||d!==i||zl.current||ja?("function"===typeof c&&(ru(n,t,c,r),i=n.memoizedState),(u=ja||au(n,t,u,r,d,i,s))?(f||"function"!==typeof o.UNSAFE_componentWillMount&&"function"!==typeof o.componentWillMount||("function"===typeof o.componentWillMount&&o.componentWillMount(),"function"===typeof o.UNSAFE_componentWillMount&&o.UNSAFE_componentWillMount()),"function"===typeof o.componentDidMount&&(n.flags|=4194308)):("function"===typeof o.componentDidMount&&(n.flags|=4194308),n.memoizedProps=r,n.memoizedState=i),o.props=r,o.state=i,o.context=s,r=u):("function"===typeof o.componentDidMount&&(n.flags|=4194308),r=!1)}else{o=n.stateNode,Ia(e,n),u=n.memoizedProps,s=n.type===n.elementType?u:tu(n.type,u),o.props=s,f=n.pendingProps,d=o.context,"object"===typeof(i=t.contextType)&&null!==i?i=La(i):i=Tl(n,i=Rl(t)?Ll:Pl.current);var p=t.getDerivedStateFromProps;(c="function"===typeof p||"function"===typeof o.getSnapshotBeforeUpdate)||"function"!==typeof o.UNSAFE_componentWillReceiveProps&&"function"!==typeof o.componentWillReceiveProps||(u!==f||d!==i)&&uu(n,o,r,i),ja=!1,d=n.memoizedState,o.state=d,Ha(n,r,o,l);var m=n.memoizedState;u!==f||d!==m||zl.current||ja?("function"===typeof p&&(ru(n,t,p,r),m=n.memoizedState),(s=ja||au(n,t,s,r,d,m,i)||!1)?(c||"function"!==typeof o.UNSAFE_componentWillUpdate&&"function"!==typeof o.componentWillUpdate||("function"===typeof o.componentWillUpdate&&o.componentWillUpdate(r,m,i),"function"===typeof o.UNSAFE_componentWillUpdate&&o.UNSAFE_componentWillUpdate(r,m,i)),"function"===typeof o.componentDidUpdate&&(n.flags|=4),"function"===typeof o.getSnapshotBeforeUpdate&&(n.flags|=1024)):("function"!==typeof o.componentDidUpdate||u===e.memoizedProps&&d===e.memoizedState||(n.flags|=4),"function"!==typeof o.getSnapshotBeforeUpdate||u===e.memoizedProps&&d===e.memoizedState||(n.flags|=1024),n.memoizedProps=r,n.memoizedState=m),o.props=r,o.state=m,o.context=i,r=s):("function"!==typeof o.componentDidUpdate||u===e.memoizedProps&&d===e.memoizedState||(n.flags|=4),"function"!==typeof o.getSnapshotBeforeUpdate||u===e.memoizedProps&&d===e.memoizedState||(n.flags|=1024),r=!1)}return Pu(e,n,t,r,a,l)}function Pu(e,n,t,r,l,a){Cu(e,n);var o=0!==(128&n.flags);if(!r&&!o)return l&&Il(n,t,!1),Bu(e,n,a);r=n.stateNode,yu.current=n;var u=o&&"function"!==typeof t.getDerivedStateFromError?null:r.render();return n.flags|=1,null!==e&&o?(n.child=ka(n,e.child,null,a),n.child=ka(n,null,u,a)):ku(e,n,u,a),n.memoizedState=r.state,l&&Il(n,t,!0),n.child}function zu(e){var n=e.stateNode;n.pendingContext?Ml(0,n.pendingContext,n.pendingContext!==n.context):n.context&&Ml(0,n.context,!1),Ya(e,n.containerInfo)}function Lu(e,n,t,r,l){return pa(),ma(l),n.flags|=256,ku(e,n,t,r),n.child}var Tu,Ru,Ou,Mu,ju={dehydrated:null,treeContext:null,retryLane:0};function Fu(e){return{baseLanes:e,cachePool:null,transitions:null}}function Iu(e,n,t){var r,l=n.pendingProps,o=Za.current,u=!1,i=0!==(128&n.flags);if((r=i)||(r=(null===e||null!==e.memoizedState)&&0!==(2&o)),r?(u=!0,n.flags&=-129):null!==e&&null===e.memoizedState||(o|=1),_l(Za,1&o),null===e)return sa(n),null!==(e=n.memoizedState)&&null!==(e=e.dehydrated)?(0===(1&n.mode)?n.lanes=1:"$!"===e.data?n.lanes=8:n.lanes=1073741824,null):(i=l.children,e=l.fallback,u?(l=n.mode,u=n.child,i={mode:"hidden",children:i},0===(1&l)&&null!==u?(u.childLanes=0,u.pendingProps=i):u=Fs(i,l,0,null),e=js(e,l,t,null),u.return=n,e.return=n,u.sibling=e,n.child=u,n.child.memoizedState=Fu(t),n.memoizedState=ju,e):Du(n,i));if(null!==(o=e.memoizedState)&&null!==(r=o.dehydrated))return function(e,n,t,r,l,o,u){if(t)return 256&n.flags?(n.flags&=-257,Uu(e,n,u,r=cu(Error(a(422))))):null!==n.memoizedState?(n.child=e.child,n.flags|=128,null):(o=r.fallback,l=n.mode,r=Fs({mode:"visible",children:r.children},l,0,null),(o=js(o,l,u,null)).flags|=2,r.return=n,o.return=n,r.sibling=o,n.child=r,0!==(1&n.mode)&&ka(n,e.child,null,u),n.child.memoizedState=Fu(u),n.memoizedState=ju,o);if(0===(1&n.mode))return Uu(e,n,u,null);if("$!"===l.data){if(r=l.nextSibling&&l.nextSibling.dataset)var i=r.dgst;return r=i,Uu(e,n,u,r=cu(o=Error(a(419)),r,void 0))}if(i=0!==(u&e.childLanes),bu||i){if(null!==(r=zi)){switch(u&-u){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}0!==(l=0!==(l&(r.suspendedLanes|u))?0:l)&&l!==o.retryLane&&(o.retryLane=l,Ma(e,l),ts(r,e,l,-1))}return hs(),Uu(e,n,u,r=cu(Error(a(421))))}return"$?"===l.data?(n.flags|=128,n.child=e.child,n=Ns.bind(null,e),l._reactRetry=n,null):(e=o.treeContext,ra=sl(l.nextSibling),ta=n,la=!0,aa=null,null!==e&&(ql[Kl++]=Xl,ql[Kl++]=Gl,ql[Kl++]=Yl,Xl=e.id,Gl=e.overflow,Yl=n),n=Du(n,r.children),n.flags|=4096,n)}(e,n,i,l,r,o,t);if(u){u=l.fallback,i=n.mode,r=(o=e.child).sibling;var s={mode:"hidden",children:l.children};return 0===(1&i)&&n.child!==o?((l=n.child).childLanes=0,l.pendingProps=s,n.deletions=null):(l=Os(o,s)).subtreeFlags=14680064&o.subtreeFlags,null!==r?u=Os(r,u):(u=js(u,i,t,null)).flags|=2,u.return=n,l.return=n,l.sibling=u,n.child=l,l=u,u=n.child,i=null===(i=e.child.memoizedState)?Fu(t):{baseLanes:i.baseLanes|t,cachePool:null,transitions:i.transitions},u.memoizedState=i,u.childLanes=e.childLanes&~t,n.memoizedState=ju,l}return e=(u=e.child).sibling,l=Os(u,{mode:"visible",children:l.children}),0===(1&n.mode)&&(l.lanes=t),l.return=n,l.sibling=null,null!==e&&(null===(t=n.deletions)?(n.deletions=[e],n.flags|=16):t.push(e)),n.child=l,n.memoizedState=null,l}function Du(e,n){return(n=Fs({mode:"visible",children:n},e.mode,0,null)).return=e,e.child=n}function Uu(e,n,t,r){return null!==r&&ma(r),ka(n,e.child,null,t),(e=Du(n,n.pendingProps.children)).flags|=2,n.memoizedState=null,e}function $u(e,n,t){e.lanes|=n;var r=e.alternate;null!==r&&(r.lanes|=n),Pa(e.return,n,t)}function Au(e,n,t,r,l){var a=e.memoizedState;null===a?e.memoizedState={isBackwards:n,rendering:null,renderingStartTime:0,last:r,tail:t,tailMode:l}:(a.isBackwards=n,a.rendering=null,a.renderingStartTime=0,a.last=r,a.tail=t,a.tailMode=l)}function Hu(e,n,t){var r=n.pendingProps,l=r.revealOrder,a=r.tail;if(ku(e,n,r.children,t),0!==(2&(r=Za.current)))r=1&r|2,n.flags|=128;else{if(null!==e&&0!==(128&e.flags))e:for(e=n.child;null!==e;){if(13===e.tag)null!==e.memoizedState&&$u(e,t,n);else if(19===e.tag)$u(e,t,n);else if(null!==e.child){e.child.return=e,e=e.child;continue}if(e===n)break e;for(;null===e.sibling;){if(null===e.return||e.return===n)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(_l(Za,r),0===(1&n.mode))n.memoizedState=null;else switch(l){case"forwards":for(t=n.child,l=null;null!==t;)null!==(e=t.alternate)&&null===eo(e)&&(l=t),t=t.sibling;null===(t=l)?(l=n.child,n.child=null):(l=t.sibling,t.sibling=null),Au(n,!1,l,t,a);break;case"backwards":for(t=null,l=n.child,n.child=null;null!==l;){if(null!==(e=l.alternate)&&null===eo(e)){n.child=l;break}e=l.sibling,l.sibling=t,t=l,l=e}Au(n,!0,t,null,a);break;case"together":Au(n,!1,null,null,void 0);break;default:n.memoizedState=null}return n.child}function Vu(e,n){0===(1&n.mode)&&null!==e&&(e.alternate=null,n.alternate=null,n.flags|=2)}function Bu(e,n,t){if(null!==e&&(n.dependencies=e.dependencies),Fi|=n.lanes,0===(t&n.childLanes))return null;if(null!==e&&n.child!==e.child)throw Error(a(153));if(null!==n.child){for(t=Os(e=n.child,e.pendingProps),n.child=t,t.return=n;null!==e.sibling;)e=e.sibling,(t=t.sibling=Os(e,e.pendingProps)).return=n;t.sibling=null}return n.child}function Wu(e,n){if(!la)switch(e.tailMode){case"hidden":n=e.tail;for(var t=null;null!==n;)null!==n.alternate&&(t=n),n=n.sibling;null===t?e.tail=null:t.sibling=null;break;case"collapsed":t=e.tail;for(var r=null;null!==t;)null!==t.alternate&&(r=t),t=t.sibling;null===r?n||null===e.tail?e.tail=null:e.tail.sibling=null:r.sibling=null}}function Qu(e){var n=null!==e.alternate&&e.alternate.child===e.child,t=0,r=0;if(n)for(var l=e.child;null!==l;)t|=l.lanes|l.childLanes,r|=14680064&l.subtreeFlags,r|=14680064&l.flags,l.return=e,l=l.sibling;else for(l=e.child;null!==l;)t|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=t,n}function qu(e,n,t){var r=n.pendingProps;switch(na(n),n.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return Qu(n),null;case 1:case 17:return Rl(n.type)&&Ol(),Qu(n),null;case 3:return r=n.stateNode,Xa(),Cl(zl),Cl(Pl),to(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),null!==e&&null!==e.child||(fa(n)?n.flags|=4:null===e||e.memoizedState.isDehydrated&&0===(256&n.flags)||(n.flags|=1024,null!==aa&&(os(aa),aa=null))),Ru(e,n),Qu(n),null;case 5:Ja(n);var l=Ka(qa.current);if(t=n.type,null!==e&&null!=n.stateNode)Ou(e,n,t,r,l),e.ref!==n.ref&&(n.flags|=512,n.flags|=2097152);else{if(!r){if(null===n.stateNode)throw Error(a(166));return Qu(n),null}if(e=Ka(Wa.current),fa(n)){r=n.stateNode,t=n.type;var o=n.memoizedProps;switch(r[dl]=n,r[pl]=o,e=0!==(1&n.mode),t){case"dialog":Ur("cancel",r),Ur("close",r);break;case"iframe":case"object":case"embed":Ur("load",r);break;case"video":case"audio":for(l=0;l<jr.length;l++)Ur(jr[l],r);break;case"source":Ur("error",r);break;case"img":case"image":case"link":Ur("error",r),Ur("load",r);break;case"details":Ur("toggle",r);break;case"input":X(r,o),Ur("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!o.multiple},Ur("invalid",r);break;case"textarea":le(r,o),Ur("invalid",r)}for(var i in ye(t,o),l=null,o)if(o.hasOwnProperty(i)){var s=o[i];"children"===i?"string"===typeof s?r.textContent!==s&&(!0!==o.suppressHydrationWarning&&Jr(r.textContent,s,e),l=["children",s]):"number"===typeof s&&r.textContent!==""+s&&(!0!==o.suppressHydrationWarning&&Jr(r.textContent,s,e),l=["children",""+s]):u.hasOwnProperty(i)&&null!=s&&"onScroll"===i&&Ur("scroll",r)}switch(t){case"input":Q(r),Z(r,o,!0);break;case"textarea":Q(r),oe(r);break;case"select":case"option":break;default:"function"===typeof o.onClick&&(r.onclick=Zr)}r=l,n.updateQueue=r,null!==r&&(n.flags|=4)}else{i=9===l.nodeType?l:l.ownerDocument,"http://www.w3.org/1999/xhtml"===e&&(e=ue(t)),"http://www.w3.org/1999/xhtml"===e?"script"===t?((e=i.createElement("div")).innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):"string"===typeof r.is?e=i.createElement(t,{is:r.is}):(e=i.createElement(t),"select"===t&&(i=e,r.multiple?i.multiple=!0:r.size&&(i.size=r.size))):e=i.createElementNS(e,t),e[dl]=n,e[pl]=r,Tu(e,n,!1,!1),n.stateNode=e;e:{switch(i=be(t,r),t){case"dialog":Ur("cancel",e),Ur("close",e),l=r;break;case"iframe":case"object":case"embed":Ur("load",e),l=r;break;case"video":case"audio":for(l=0;l<jr.length;l++)Ur(jr[l],e);l=r;break;case"source":Ur("error",e),l=r;break;case"img":case"image":case"link":Ur("error",e),Ur("load",e),l=r;break;case"details":Ur("toggle",e),l=r;break;case"input":X(e,r),l=Y(e,r),Ur("invalid",e);break;case"option":default:l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=I({},r,{value:void 0}),Ur("invalid",e);break;case"textarea":le(e,r),l=re(e,r),Ur("invalid",e)}for(o in ye(t,l),s=l)if(s.hasOwnProperty(o)){var c=s[o];"style"===o?ve(e,c):"dangerouslySetInnerHTML"===o?null!=(c=c?c.__html:void 0)&&fe(e,c):"children"===o?"string"===typeof c?("textarea"!==t||""!==c)&&de(e,c):"number"===typeof c&&de(e,""+c):"suppressContentEditableWarning"!==o&&"suppressHydrationWarning"!==o&&"autoFocus"!==o&&(u.hasOwnProperty(o)?null!=c&&"onScroll"===o&&Ur("scroll",e):null!=c&&b(e,o,c,i))}switch(t){case"input":Q(e),Z(e,r,!1);break;case"textarea":Q(e),oe(e);break;case"option":null!=r.value&&e.setAttribute("value",""+B(r.value));break;case"select":e.multiple=!!r.multiple,null!=(o=r.value)?te(e,!!r.multiple,o,!1):null!=r.defaultValue&&te(e,!!r.multiple,r.defaultValue,!0);break;default:"function"===typeof l.onClick&&(e.onclick=Zr)}switch(t){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(n.flags|=4)}null!==n.ref&&(n.flags|=512,n.flags|=2097152)}return Qu(n),null;case 6:if(e&&null!=n.stateNode)Mu(e,n,e.memoizedProps,r);else{if("string"!==typeof r&&null===n.stateNode)throw Error(a(166));if(t=Ka(qa.current),Ka(Wa.current),fa(n)){if(r=n.stateNode,t=n.memoizedProps,r[dl]=n,(o=r.nodeValue!==t)&&null!==(e=ta))switch(e.tag){case 3:Jr(r.nodeValue,t,0!==(1&e.mode));break;case 5:!0!==e.memoizedProps.suppressHydrationWarning&&Jr(r.nodeValue,t,0!==(1&e.mode))}o&&(n.flags|=4)}else(r=(9===t.nodeType?t:t.ownerDocument).createTextNode(r))[dl]=n,n.stateNode=r}return Qu(n),null;case 13:if(Cl(Za),r=n.memoizedState,null===e||null!==e.memoizedState&&null!==e.memoizedState.dehydrated){if(la&&null!==ra&&0!==(1&n.mode)&&0===(128&n.flags))da(),pa(),n.flags|=98560,o=!1;else if(o=fa(n),null!==r&&null!==r.dehydrated){if(null===e){if(!o)throw Error(a(318));if(!(o=null!==(o=n.memoizedState)?o.dehydrated:null))throw Error(a(317));o[dl]=n}else pa(),0===(128&n.flags)&&(n.memoizedState=null),n.flags|=4;Qu(n),o=!1}else null!==aa&&(os(aa),aa=null),o=!0;if(!o)return 65536&n.flags?n:null}return 0!==(128&n.flags)?(n.lanes=t,n):((r=null!==r)!==(null!==e&&null!==e.memoizedState)&&r&&(n.child.flags|=8192,0!==(1&n.mode)&&(null===e||0!==(1&Za.current)?0===Mi&&(Mi=3):hs())),null!==n.updateQueue&&(n.flags|=4),Qu(n),null);case 4:return Xa(),Ru(e,n),null===e&&Hr(n.stateNode.containerInfo),Qu(n),null;case 10:return Na(n.type._context),Qu(n),null;case 19:if(Cl(Za),null===(o=n.memoizedState))return Qu(n),null;if(r=0!==(128&n.flags),null===(i=o.rendering))if(r)Wu(o,!1);else{if(0!==Mi||null!==e&&0!==(128&e.flags))for(e=n.child;null!==e;){if(null!==(i=eo(e))){for(n.flags|=128,Wu(o,!1),null!==(r=i.updateQueue)&&(n.updateQueue=r,n.flags|=4),n.subtreeFlags=0,r=t,t=n.child;null!==t;)e=r,(o=t).flags&=14680066,null===(i=o.alternate)?(o.childLanes=0,o.lanes=e,o.child=null,o.subtreeFlags=0,o.memoizedProps=null,o.memoizedState=null,o.updateQueue=null,o.dependencies=null,o.stateNode=null):(o.childLanes=i.childLanes,o.lanes=i.lanes,o.child=i.child,o.subtreeFlags=0,o.deletions=null,o.memoizedProps=i.memoizedProps,o.memoizedState=i.memoizedState,o.updateQueue=i.updateQueue,o.type=i.type,e=i.dependencies,o.dependencies=null===e?null:{lanes:e.lanes,firstContext:e.firstContext}),t=t.sibling;return _l(Za,1&Za.current|2),n.child}e=e.sibling}null!==o.tail&&Ge()>Hi&&(n.flags|=128,r=!0,Wu(o,!1),n.lanes=4194304)}else{if(!r)if(null!==(e=eo(i))){if(n.flags|=128,r=!0,null!==(t=e.updateQueue)&&(n.updateQueue=t,n.flags|=4),Wu(o,!0),null===o.tail&&"hidden"===o.tailMode&&!i.alternate&&!la)return Qu(n),null}else 2*Ge()-o.renderingStartTime>Hi&&1073741824!==t&&(n.flags|=128,r=!0,Wu(o,!1),n.lanes=4194304);o.isBackwards?(i.sibling=n.child,n.child=i):(null!==(t=o.last)?t.sibling=i:n.child=i,o.last=i)}return null!==o.tail?(n=o.tail,o.rendering=n,o.tail=n.sibling,o.renderingStartTime=Ge(),n.sibling=null,t=Za.current,_l(Za,r?1&t|2:1&t),n):(Qu(n),null);case 22:case 23:return fs(),r=null!==n.memoizedState,null!==e&&null!==e.memoizedState!==r&&(n.flags|=8192),r&&0!==(1&n.mode)?0!==(1073741824&Ri)&&(Qu(n),6&n.subtreeFlags&&(n.flags|=8192)):Qu(n),null;case 24:case 25:return null}throw Error(a(156,n.tag))}function Ku(e,n){switch(na(n),n.tag){case 1:return Rl(n.type)&&Ol(),65536&(e=n.flags)?(n.flags=-65537&e|128,n):null;case 3:return Xa(),Cl(zl),Cl(Pl),to(),0!==(65536&(e=n.flags))&&0===(128&e)?(n.flags=-65537&e|128,n):null;case 5:return Ja(n),null;case 13:if(Cl(Za),null!==(e=n.memoizedState)&&null!==e.dehydrated){if(null===n.alternate)throw Error(a(340));pa()}return 65536&(e=n.flags)?(n.flags=-65537&e|128,n):null;case 19:return Cl(Za),null;case 4:return Xa(),null;case 10:return Na(n.type._context),null;case 22:case 23:return fs(),null;default:return null}}Tu=function(e,n){for(var t=n.child;null!==t;){if(5===t.tag||6===t.tag)e.appendChild(t.stateNode);else if(4!==t.tag&&null!==t.child){t.child.return=t,t=t.child;continue}if(t===n)break;for(;null===t.sibling;){if(null===t.return||t.return===n)return;t=t.return}t.sibling.return=t.return,t=t.sibling}},Ru=function(){},Ou=function(e,n,t,r){var l=e.memoizedProps;if(l!==r){e=n.stateNode,Ka(Wa.current);var a,o=null;switch(t){case"input":l=Y(e,l),r=Y(e,r),o=[];break;case"select":l=I({},l,{value:void 0}),r=I({},r,{value:void 0}),o=[];break;case"textarea":l=re(e,l),r=re(e,r),o=[];break;default:"function"!==typeof l.onClick&&"function"===typeof r.onClick&&(e.onclick=Zr)}for(c in ye(t,r),t=null,l)if(!r.hasOwnProperty(c)&&l.hasOwnProperty(c)&&null!=l[c])if("style"===c){var i=l[c];for(a in i)i.hasOwnProperty(a)&&(t||(t={}),t[a]="")}else"dangerouslySetInnerHTML"!==c&&"children"!==c&&"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&"autoFocus"!==c&&(u.hasOwnProperty(c)?o||(o=[]):(o=o||[]).push(c,null));for(c in r){var s=r[c];if(i=null!=l?l[c]:void 0,r.hasOwnProperty(c)&&s!==i&&(null!=s||null!=i))if("style"===c)if(i){for(a in i)!i.hasOwnProperty(a)||s&&s.hasOwnProperty(a)||(t||(t={}),t[a]="");for(a in s)s.hasOwnProperty(a)&&i[a]!==s[a]&&(t||(t={}),t[a]=s[a])}else t||(o||(o=[]),o.push(c,t)),t=s;else"dangerouslySetInnerHTML"===c?(s=s?s.__html:void 0,i=i?i.__html:void 0,null!=s&&i!==s&&(o=o||[]).push(c,s)):"children"===c?"string"!==typeof s&&"number"!==typeof s||(o=o||[]).push(c,""+s):"suppressContentEditableWarning"!==c&&"suppressHydrationWarning"!==c&&(u.hasOwnProperty(c)?(null!=s&&"onScroll"===c&&Ur("scroll",e),o||i===s||(o=[])):(o=o||[]).push(c,s))}t&&(o=o||[]).push("style",t);var c=o;(n.updateQueue=c)&&(n.flags|=4)}},Mu=function(e,n,t,r){t!==r&&(n.flags|=4)};var Yu=!1,Xu=!1,Gu="function"===typeof WeakSet?WeakSet:Set,Ju=null;function Zu(e,n){var t=e.ref;if(null!==t)if("function"===typeof t)try{t(null)}catch(r){Es(e,n,r)}else t.current=null}function ei(e,n,t){try{t()}catch(r){Es(e,n,r)}}var ni=!1;function ti(e,n,t){var r=n.updateQueue;if(null!==(r=null!==r?r.lastEffect:null)){var l=r=r.next;do{if((l.tag&e)===e){var a=l.destroy;l.destroy=void 0,void 0!==a&&ei(n,t,a)}l=l.next}while(l!==r)}}function ri(e,n){if(null!==(n=null!==(n=n.updateQueue)?n.lastEffect:null)){var t=n=n.next;do{if((t.tag&e)===e){var r=t.create;t.destroy=r()}t=t.next}while(t!==n)}}function li(e){var n=e.ref;if(null!==n){var t=e.stateNode;e.tag,e=t,"function"===typeof n?n(e):n.current=e}}function ai(e){var n=e.alternate;null!==n&&(e.alternate=null,ai(n)),e.child=null,e.deletions=null,e.sibling=null,5===e.tag&&(null!==(n=e.stateNode)&&(delete n[dl],delete n[pl],delete n[hl],delete n[vl],delete n[gl])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function oi(e){return 5===e.tag||3===e.tag||4===e.tag}function ui(e){e:for(;;){for(;null===e.sibling;){if(null===e.return||oi(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;5!==e.tag&&6!==e.tag&&18!==e.tag;){if(2&e.flags)continue e;if(null===e.child||4===e.tag)continue e;e.child.return=e,e=e.child}if(!(2&e.flags))return e.stateNode}}function ii(e,n,t){var r=e.tag;if(5===r||6===r)e=e.stateNode,n?8===t.nodeType?t.parentNode.insertBefore(e,n):t.insertBefore(e,n):(8===t.nodeType?(n=t.parentNode).insertBefore(e,t):(n=t).appendChild(e),null!==(t=t._reactRootContainer)&&void 0!==t||null!==n.onclick||(n.onclick=Zr));else if(4!==r&&null!==(e=e.child))for(ii(e,n,t),e=e.sibling;null!==e;)ii(e,n,t),e=e.sibling}function si(e,n,t){var r=e.tag;if(5===r||6===r)e=e.stateNode,n?t.insertBefore(e,n):t.appendChild(e);else if(4!==r&&null!==(e=e.child))for(si(e,n,t),e=e.sibling;null!==e;)si(e,n,t),e=e.sibling}var ci=null,fi=!1;function di(e,n,t){for(t=t.child;null!==t;)pi(e,n,t),t=t.sibling}function pi(e,n,t){if(an&&"function"===typeof an.onCommitFiberUnmount)try{an.onCommitFiberUnmount(ln,t)}catch(u){}switch(t.tag){case 5:Xu||Zu(t,n);case 6:var r=ci,l=fi;ci=null,di(e,n,t),fi=l,null!==(ci=r)&&(fi?(e=ci,t=t.stateNode,8===e.nodeType?e.parentNode.removeChild(t):e.removeChild(t)):ci.removeChild(t.stateNode));break;case 18:null!==ci&&(fi?(e=ci,t=t.stateNode,8===e.nodeType?il(e.parentNode,t):1===e.nodeType&&il(e,t),Vn(e)):il(ci,t.stateNode));break;case 4:r=ci,l=fi,ci=t.stateNode.containerInfo,fi=!0,di(e,n,t),ci=r,fi=l;break;case 0:case 11:case 14:case 15:if(!Xu&&(null!==(r=t.updateQueue)&&null!==(r=r.lastEffect))){l=r=r.next;do{var a=l,o=a.destroy;a=a.tag,void 0!==o&&(0!==(2&a)||0!==(4&a))&&ei(t,n,o),l=l.next}while(l!==r)}di(e,n,t);break;case 1:if(!Xu&&(Zu(t,n),"function"===typeof(r=t.stateNode).componentWillUnmount))try{r.props=t.memoizedProps,r.state=t.memoizedState,r.componentWillUnmount()}catch(u){Es(t,n,u)}di(e,n,t);break;case 21:di(e,n,t);break;case 22:1&t.mode?(Xu=(r=Xu)||null!==t.memoizedState,di(e,n,t),Xu=r):di(e,n,t);break;default:di(e,n,t)}}function mi(e){var n=e.updateQueue;if(null!==n){e.updateQueue=null;var t=e.stateNode;null===t&&(t=e.stateNode=new Gu),n.forEach((function(n){var r=Ps.bind(null,e,n);t.has(n)||(t.add(n),n.then(r,r))}))}}function hi(e,n){var t=n.deletions;if(null!==t)for(var r=0;r<t.length;r++){var l=t[r];try{var o=e,u=n,i=u;e:for(;null!==i;){switch(i.tag){case 5:ci=i.stateNode,fi=!1;break e;case 3:case 4:ci=i.stateNode.containerInfo,fi=!0;break e}i=i.return}if(null===ci)throw Error(a(160));pi(o,u,l),ci=null,fi=!1;var s=l.alternate;null!==s&&(s.return=null),l.return=null}catch(c){Es(l,n,c)}}if(12854&n.subtreeFlags)for(n=n.child;null!==n;)vi(n,e),n=n.sibling}function vi(e,n){var t=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(hi(n,e),gi(e),4&r){try{ti(3,e,e.return),ri(3,e)}catch(v){Es(e,e.return,v)}try{ti(5,e,e.return)}catch(v){Es(e,e.return,v)}}break;case 1:hi(n,e),gi(e),512&r&&null!==t&&Zu(t,t.return);break;case 5:if(hi(n,e),gi(e),512&r&&null!==t&&Zu(t,t.return),32&e.flags){var l=e.stateNode;try{de(l,"")}catch(v){Es(e,e.return,v)}}if(4&r&&null!=(l=e.stateNode)){var o=e.memoizedProps,u=null!==t?t.memoizedProps:o,i=e.type,s=e.updateQueue;if(e.updateQueue=null,null!==s)try{"input"===i&&"radio"===o.type&&null!=o.name&&G(l,o),be(i,u);var c=be(i,o);for(u=0;u<s.length;u+=2){var f=s[u],d=s[u+1];"style"===f?ve(l,d):"dangerouslySetInnerHTML"===f?fe(l,d):"children"===f?de(l,d):b(l,f,d,c)}switch(i){case"input":J(l,o);break;case"textarea":ae(l,o);break;case"select":var p=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!o.multiple;var m=o.value;null!=m?te(l,!!o.multiple,m,!1):p!==!!o.multiple&&(null!=o.defaultValue?te(l,!!o.multiple,o.defaultValue,!0):te(l,!!o.multiple,o.multiple?[]:"",!1))}l[pl]=o}catch(v){Es(e,e.return,v)}}break;case 6:if(hi(n,e),gi(e),4&r){if(null===e.stateNode)throw Error(a(162));l=e.stateNode,o=e.memoizedProps;try{l.nodeValue=o}catch(v){Es(e,e.return,v)}}break;case 3:if(hi(n,e),gi(e),4&r&&null!==t&&t.memoizedState.isDehydrated)try{Vn(n.containerInfo)}catch(v){Es(e,e.return,v)}break;case 4:default:hi(n,e),gi(e);break;case 13:hi(n,e),gi(e),8192&(l=e.child).flags&&(o=null!==l.memoizedState,l.stateNode.isHidden=o,!o||null!==l.alternate&&null!==l.alternate.memoizedState||(Ai=Ge())),4&r&&mi(e);break;case 22:if(f=null!==t&&null!==t.memoizedState,1&e.mode?(Xu=(c=Xu)||f,hi(n,e),Xu=c):hi(n,e),gi(e),8192&r){if(c=null!==e.memoizedState,(e.stateNode.isHidden=c)&&!f&&0!==(1&e.mode))for(Ju=e,f=e.child;null!==f;){for(d=Ju=f;null!==Ju;){switch(m=(p=Ju).child,p.tag){case 0:case 11:case 14:case 15:ti(4,p,p.return);break;case 1:Zu(p,p.return);var h=p.stateNode;if("function"===typeof h.componentWillUnmount){r=p,t=p.return;try{n=r,h.props=n.memoizedProps,h.state=n.memoizedState,h.componentWillUnmount()}catch(v){Es(r,t,v)}}break;case 5:Zu(p,p.return);break;case 22:if(null!==p.memoizedState){wi(d);continue}}null!==m?(m.return=p,Ju=m):wi(d)}f=f.sibling}e:for(f=null,d=e;;){if(5===d.tag){if(null===f){f=d;try{l=d.stateNode,c?"function"===typeof(o=l.style).setProperty?o.setProperty("display","none","important"):o.display="none":(i=d.stateNode,u=void 0!==(s=d.memoizedProps.style)&&null!==s&&s.hasOwnProperty("display")?s.display:null,i.style.display=he("display",u))}catch(v){Es(e,e.return,v)}}}else if(6===d.tag){if(null===f)try{d.stateNode.nodeValue=c?"":d.memoizedProps}catch(v){Es(e,e.return,v)}}else if((22!==d.tag&&23!==d.tag||null===d.memoizedState||d===e)&&null!==d.child){d.child.return=d,d=d.child;continue}if(d===e)break e;for(;null===d.sibling;){if(null===d.return||d.return===e)break e;f===d&&(f=null),d=d.return}f===d&&(f=null),d.sibling.return=d.return,d=d.sibling}}break;case 19:hi(n,e),gi(e),4&r&&mi(e);case 21:}}function gi(e){var n=e.flags;if(2&n){try{e:{for(var t=e.return;null!==t;){if(oi(t)){var r=t;break e}t=t.return}throw Error(a(160))}switch(r.tag){case 5:var l=r.stateNode;32&r.flags&&(de(l,""),r.flags&=-33),si(e,ui(e),l);break;case 3:case 4:var o=r.stateNode.containerInfo;ii(e,ui(e),o);break;default:throw Error(a(161))}}catch(u){Es(e,e.return,u)}e.flags&=-3}4096&n&&(e.flags&=-4097)}function yi(e,n,t){Ju=e,bi(e,n,t)}function bi(e,n,t){for(var r=0!==(1&e.mode);null!==Ju;){var l=Ju,a=l.child;if(22===l.tag&&r){var o=null!==l.memoizedState||Yu;if(!o){var u=l.alternate,i=null!==u&&null!==u.memoizedState||Xu;u=Yu;var s=Xu;if(Yu=o,(Xu=i)&&!s)for(Ju=l;null!==Ju;)i=(o=Ju).child,22===o.tag&&null!==o.memoizedState?Si(l):null!==i?(i.return=o,Ju=i):Si(l);for(;null!==a;)Ju=a,bi(a,n,t),a=a.sibling;Ju=l,Yu=u,Xu=s}ki(e)}else 0!==(8772&l.subtreeFlags)&&null!==a?(a.return=l,Ju=a):ki(e)}}function ki(e){for(;null!==Ju;){var n=Ju;if(0!==(8772&n.flags)){var t=n.alternate;try{if(0!==(8772&n.flags))switch(n.tag){case 0:case 11:case 15:Xu||ri(5,n);break;case 1:var r=n.stateNode;if(4&n.flags&&!Xu)if(null===t)r.componentDidMount();else{var l=n.elementType===n.type?t.memoizedProps:tu(n.type,t.memoizedProps);r.componentDidUpdate(l,t.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var o=n.updateQueue;null!==o&&Va(n,o,r);break;case 3:var u=n.updateQueue;if(null!==u){if(t=null,null!==n.child)switch(n.child.tag){case 5:case 1:t=n.child.stateNode}Va(n,u,t)}break;case 5:var i=n.stateNode;if(null===t&&4&n.flags){t=i;var s=n.memoizedProps;switch(n.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&t.focus();break;case"img":s.src&&(t.src=s.src)}}break;case 6:case 4:case 12:case 19:case 17:case 21:case 22:case 23:case 25:break;case 13:if(null===n.memoizedState){var c=n.alternate;if(null!==c){var f=c.memoizedState;if(null!==f){var d=f.dehydrated;null!==d&&Vn(d)}}}break;default:throw Error(a(163))}Xu||512&n.flags&&li(n)}catch(p){Es(n,n.return,p)}}if(n===e){Ju=null;break}if(null!==(t=n.sibling)){t.return=n.return,Ju=t;break}Ju=n.return}}function wi(e){for(;null!==Ju;){var n=Ju;if(n===e){Ju=null;break}var t=n.sibling;if(null!==t){t.return=n.return,Ju=t;break}Ju=n.return}}function Si(e){for(;null!==Ju;){var n=Ju;try{switch(n.tag){case 0:case 11:case 15:var t=n.return;try{ri(4,n)}catch(i){Es(n,t,i)}break;case 1:var r=n.stateNode;if("function"===typeof r.componentDidMount){var l=n.return;try{r.componentDidMount()}catch(i){Es(n,l,i)}}var a=n.return;try{li(n)}catch(i){Es(n,a,i)}break;case 5:var o=n.return;try{li(n)}catch(i){Es(n,o,i)}}}catch(i){Es(n,n.return,i)}if(n===e){Ju=null;break}var u=n.sibling;if(null!==u){u.return=n.return,Ju=u;break}Ju=n.return}}var xi,Ei=Math.ceil,Ci=k.ReactCurrentDispatcher,_i=k.ReactCurrentOwner,Ni=k.ReactCurrentBatchConfig,Pi=0,zi=null,Li=null,Ti=0,Ri=0,Oi=El(0),Mi=0,ji=null,Fi=0,Ii=0,Di=0,Ui=null,$i=null,Ai=0,Hi=1/0,Vi=null,Bi=!1,Wi=null,Qi=null,qi=!1,Ki=null,Yi=0,Xi=0,Gi=null,Ji=-1,Zi=0;function es(){return 0!==(6&Pi)?Ge():-1!==Ji?Ji:Ji=Ge()}function ns(e){return 0===(1&e.mode)?1:0!==(2&Pi)&&0!==Ti?Ti&-Ti:null!==ha.transition?(0===Zi&&(Zi=vn()),Zi):0!==(e=kn)?e:e=void 0===(e=window.event)?16:Gn(e.type)}function ts(e,n,t,r){if(50<Xi)throw Xi=0,Gi=null,Error(a(185));yn(e,t,r),0!==(2&Pi)&&e===zi||(e===zi&&(0===(2&Pi)&&(Ii|=t),4===Mi&&us(e,Ti)),rs(e,r),1===t&&0===Pi&&0===(1&n.mode)&&(Hi=Ge()+500,Ul&&Hl()))}function rs(e,n){var t=e.callbackNode;!function(e,n){for(var t=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,a=e.pendingLanes;0<a;){var o=31-on(a),u=1<<o,i=l[o];-1===i?0!==(u&t)&&0===(u&r)||(l[o]=mn(u,n)):i<=n&&(e.expiredLanes|=u),a&=~u}}(e,n);var r=pn(e,e===zi?Ti:0);if(0===r)null!==t&&Ke(t),e.callbackNode=null,e.callbackPriority=0;else if(n=r&-r,e.callbackPriority!==n){if(null!=t&&Ke(t),1===n)0===e.tag?function(e){Ul=!0,Al(e)}(is.bind(null,e)):Al(is.bind(null,e)),ol((function(){0===(6&Pi)&&Hl()})),t=null;else{switch(wn(r)){case 1:t=Ze;break;case 4:t=en;break;case 16:default:t=nn;break;case 536870912:t=rn}t=zs(t,ls.bind(null,e))}e.callbackPriority=n,e.callbackNode=t}}function ls(e,n){if(Ji=-1,Zi=0,0!==(6&Pi))throw Error(a(327));var t=e.callbackNode;if(Ss()&&e.callbackNode!==t)return null;var r=pn(e,e===zi?Ti:0);if(0===r)return null;if(0!==(30&r)||0!==(r&e.expiredLanes)||n)n=vs(e,r);else{n=r;var l=Pi;Pi|=2;var o=ms();for(zi===e&&Ti===n||(Vi=null,Hi=Ge()+500,ds(e,n));;)try{ys();break}catch(i){ps(e,i)}_a(),Ci.current=o,Pi=l,null!==Li?n=0:(zi=null,Ti=0,n=Mi)}if(0!==n){if(2===n&&(0!==(l=hn(e))&&(r=l,n=as(e,l))),1===n)throw t=ji,ds(e,0),us(e,r),rs(e,Ge()),t;if(6===n)us(e,r);else{if(l=e.current.alternate,0===(30&r)&&!function(e){for(var n=e;;){if(16384&n.flags){var t=n.updateQueue;if(null!==t&&null!==(t=t.stores))for(var r=0;r<t.length;r++){var l=t[r],a=l.getSnapshot;l=l.value;try{if(!ur(a(),l))return!1}catch(u){return!1}}}if(t=n.child,16384&n.subtreeFlags&&null!==t)t.return=n,n=t;else{if(n===e)break;for(;null===n.sibling;){if(null===n.return||n.return===e)return!0;n=n.return}n.sibling.return=n.return,n=n.sibling}}return!0}(l)&&(2===(n=vs(e,r))&&(0!==(o=hn(e))&&(r=o,n=as(e,o))),1===n))throw t=ji,ds(e,0),us(e,r),rs(e,Ge()),t;switch(e.finishedWork=l,e.finishedLanes=r,n){case 0:case 1:throw Error(a(345));case 2:case 5:ws(e,$i,Vi);break;case 3:if(us(e,r),(130023424&r)===r&&10<(n=Ai+500-Ge())){if(0!==pn(e,0))break;if(((l=e.suspendedLanes)&r)!==r){es(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=rl(ws.bind(null,e,$i,Vi),n);break}ws(e,$i,Vi);break;case 4:if(us(e,r),(4194240&r)===r)break;for(n=e.eventTimes,l=-1;0<r;){var u=31-on(r);o=1<<u,(u=n[u])>l&&(l=u),r&=~o}if(r=l,10<(r=(120>(r=Ge()-r)?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Ei(r/1960))-r)){e.timeoutHandle=rl(ws.bind(null,e,$i,Vi),r);break}ws(e,$i,Vi);break;default:throw Error(a(329))}}}return rs(e,Ge()),e.callbackNode===t?ls.bind(null,e):null}function as(e,n){var t=Ui;return e.current.memoizedState.isDehydrated&&(ds(e,n).flags|=256),2!==(e=vs(e,n))&&(n=$i,$i=t,null!==n&&os(n)),e}function os(e){null===$i?$i=e:$i.push.apply($i,e)}function us(e,n){for(n&=~Di,n&=~Ii,e.suspendedLanes|=n,e.pingedLanes&=~n,e=e.expirationTimes;0<n;){var t=31-on(n),r=1<<t;e[t]=-1,n&=~r}}function is(e){if(0!==(6&Pi))throw Error(a(327));Ss();var n=pn(e,0);if(0===(1&n))return rs(e,Ge()),null;var t=vs(e,n);if(0!==e.tag&&2===t){var r=hn(e);0!==r&&(n=r,t=as(e,r))}if(1===t)throw t=ji,ds(e,0),us(e,n),rs(e,Ge()),t;if(6===t)throw Error(a(345));return e.finishedWork=e.current.alternate,e.finishedLanes=n,ws(e,$i,Vi),rs(e,Ge()),null}function ss(e,n){var t=Pi;Pi|=1;try{return e(n)}finally{0===(Pi=t)&&(Hi=Ge()+500,Ul&&Hl())}}function cs(e){null!==Ki&&0===Ki.tag&&0===(6&Pi)&&Ss();var n=Pi;Pi|=1;var t=Ni.transition,r=kn;try{if(Ni.transition=null,kn=1,e)return e()}finally{kn=r,Ni.transition=t,0===(6&(Pi=n))&&Hl()}}function fs(){Ri=Oi.current,Cl(Oi)}function ds(e,n){e.finishedWork=null,e.finishedLanes=0;var t=e.timeoutHandle;if(-1!==t&&(e.timeoutHandle=-1,ll(t)),null!==Li)for(t=Li.return;null!==t;){var r=t;switch(na(r),r.tag){case 1:null!==(r=r.type.childContextTypes)&&void 0!==r&&Ol();break;case 3:Xa(),Cl(zl),Cl(Pl),to();break;case 5:Ja(r);break;case 4:Xa();break;case 13:case 19:Cl(Za);break;case 10:Na(r.type._context);break;case 22:case 23:fs()}t=t.return}if(zi=e,Li=e=Os(e.current,null),Ti=Ri=n,Mi=0,ji=null,Di=Ii=Fi=0,$i=Ui=null,null!==Ta){for(n=0;n<Ta.length;n++)if(null!==(r=(t=Ta[n]).interleaved)){t.interleaved=null;var l=r.next,a=t.pending;if(null!==a){var o=a.next;a.next=l,r.next=o}t.pending=r}Ta=null}return e}function ps(e,n){for(;;){var t=Li;try{if(_a(),ro.current=Jo,so){for(var r=oo.memoizedState;null!==r;){var l=r.queue;null!==l&&(l.pending=null),r=r.next}so=!1}if(ao=0,io=uo=oo=null,co=!1,fo=0,_i.current=null,null===t||null===t.return){Mi=1,ji=n,Li=null;break}e:{var o=e,u=t.return,i=t,s=n;if(n=Ti,i.flags|=32768,null!==s&&"object"===typeof s&&"function"===typeof s.then){var c=s,f=i,d=f.tag;if(0===(1&f.mode)&&(0===d||11===d||15===d)){var p=f.alternate;p?(f.updateQueue=p.updateQueue,f.memoizedState=p.memoizedState,f.lanes=p.lanes):(f.updateQueue=null,f.memoizedState=null)}var m=vu(u);if(null!==m){m.flags&=-257,gu(m,u,i,0,n),1&m.mode&&hu(o,c,n),s=c;var h=(n=m).updateQueue;if(null===h){var v=new Set;v.add(s),n.updateQueue=v}else h.add(s);break e}if(0===(1&n)){hu(o,c,n),hs();break e}s=Error(a(426))}else if(la&&1&i.mode){var g=vu(u);if(null!==g){0===(65536&g.flags)&&(g.flags|=256),gu(g,u,i,0,n),ma(su(s,i));break e}}o=s=su(s,i),4!==Mi&&(Mi=2),null===Ui?Ui=[o]:Ui.push(o),o=u;do{switch(o.tag){case 3:o.flags|=65536,n&=-n,o.lanes|=n,Aa(o,pu(0,s,n));break e;case 1:i=s;var y=o.type,b=o.stateNode;if(0===(128&o.flags)&&("function"===typeof y.getDerivedStateFromError||null!==b&&"function"===typeof b.componentDidCatch&&(null===Qi||!Qi.has(b)))){o.flags|=65536,n&=-n,o.lanes|=n,Aa(o,mu(o,i,n));break e}}o=o.return}while(null!==o)}ks(t)}catch(k){n=k,Li===t&&null!==t&&(Li=t=t.return);continue}break}}function ms(){var e=Ci.current;return Ci.current=Jo,null===e?Jo:e}function hs(){0!==Mi&&3!==Mi&&2!==Mi||(Mi=4),null===zi||0===(268435455&Fi)&&0===(268435455&Ii)||us(zi,Ti)}function vs(e,n){var t=Pi;Pi|=2;var r=ms();for(zi===e&&Ti===n||(Vi=null,ds(e,n));;)try{gs();break}catch(l){ps(e,l)}if(_a(),Pi=t,Ci.current=r,null!==Li)throw Error(a(261));return zi=null,Ti=0,Mi}function gs(){for(;null!==Li;)bs(Li)}function ys(){for(;null!==Li&&!Ye();)bs(Li)}function bs(e){var n=xi(e.alternate,e,Ri);e.memoizedProps=e.pendingProps,null===n?ks(e):Li=n,_i.current=null}function ks(e){var n=e;do{var t=n.alternate;if(e=n.return,0===(32768&n.flags)){if(null!==(t=qu(t,n,Ri)))return void(Li=t)}else{if(null!==(t=Ku(t,n)))return t.flags&=32767,void(Li=t);if(null===e)return Mi=6,void(Li=null);e.flags|=32768,e.subtreeFlags=0,e.deletions=null}if(null!==(n=n.sibling))return void(Li=n);Li=n=e}while(null!==n);0===Mi&&(Mi=5)}function ws(e,n,t){var r=kn,l=Ni.transition;try{Ni.transition=null,kn=1,function(e,n,t,r){do{Ss()}while(null!==Ki);if(0!==(6&Pi))throw Error(a(327));t=e.finishedWork;var l=e.finishedLanes;if(null===t)return null;if(e.finishedWork=null,e.finishedLanes=0,t===e.current)throw Error(a(177));e.callbackNode=null,e.callbackPriority=0;var o=t.lanes|t.childLanes;if(function(e,n){var t=e.pendingLanes&~n;e.pendingLanes=n,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=n,e.mutableReadLanes&=n,e.entangledLanes&=n,n=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<t;){var l=31-on(t),a=1<<l;n[l]=0,r[l]=-1,e[l]=-1,t&=~a}}(e,o),e===zi&&(Li=zi=null,Ti=0),0===(2064&t.subtreeFlags)&&0===(2064&t.flags)||qi||(qi=!0,zs(nn,(function(){return Ss(),null}))),o=0!==(15990&t.flags),0!==(15990&t.subtreeFlags)||o){o=Ni.transition,Ni.transition=null;var u=kn;kn=1;var i=Pi;Pi|=4,_i.current=null,function(e,n){if(el=Wn,pr(e=dr())){if("selectionStart"in e)var t={start:e.selectionStart,end:e.selectionEnd};else e:{var r=(t=(t=e.ownerDocument)&&t.defaultView||window).getSelection&&t.getSelection();if(r&&0!==r.rangeCount){t=r.anchorNode;var l=r.anchorOffset,o=r.focusNode;r=r.focusOffset;try{t.nodeType,o.nodeType}catch(w){t=null;break e}var u=0,i=-1,s=-1,c=0,f=0,d=e,p=null;n:for(;;){for(var m;d!==t||0!==l&&3!==d.nodeType||(i=u+l),d!==o||0!==r&&3!==d.nodeType||(s=u+r),3===d.nodeType&&(u+=d.nodeValue.length),null!==(m=d.firstChild);)p=d,d=m;for(;;){if(d===e)break n;if(p===t&&++c===l&&(i=u),p===o&&++f===r&&(s=u),null!==(m=d.nextSibling))break;p=(d=p).parentNode}d=m}t=-1===i||-1===s?null:{start:i,end:s}}else t=null}t=t||{start:0,end:0}}else t=null;for(nl={focusedElem:e,selectionRange:t},Wn=!1,Ju=n;null!==Ju;)if(e=(n=Ju).child,0!==(1028&n.subtreeFlags)&&null!==e)e.return=n,Ju=e;else for(;null!==Ju;){n=Ju;try{var h=n.alternate;if(0!==(1024&n.flags))switch(n.tag){case 0:case 11:case 15:case 5:case 6:case 4:case 17:break;case 1:if(null!==h){var v=h.memoizedProps,g=h.memoizedState,y=n.stateNode,b=y.getSnapshotBeforeUpdate(n.elementType===n.type?v:tu(n.type,v),g);y.__reactInternalSnapshotBeforeUpdate=b}break;case 3:var k=n.stateNode.containerInfo;1===k.nodeType?k.textContent="":9===k.nodeType&&k.documentElement&&k.removeChild(k.documentElement);break;default:throw Error(a(163))}}catch(w){Es(n,n.return,w)}if(null!==(e=n.sibling)){e.return=n.return,Ju=e;break}Ju=n.return}h=ni,ni=!1}(e,t),vi(t,e),mr(nl),Wn=!!el,nl=el=null,e.current=t,yi(t,e,l),Xe(),Pi=i,kn=u,Ni.transition=o}else e.current=t;if(qi&&(qi=!1,Ki=e,Yi=l),o=e.pendingLanes,0===o&&(Qi=null),function(e){if(an&&"function"===typeof an.onCommitFiberRoot)try{an.onCommitFiberRoot(ln,e,void 0,128===(128&e.current.flags))}catch(n){}}(t.stateNode),rs(e,Ge()),null!==n)for(r=e.onRecoverableError,t=0;t<n.length;t++)l=n[t],r(l.value,{componentStack:l.stack,digest:l.digest});if(Bi)throw Bi=!1,e=Wi,Wi=null,e;0!==(1&Yi)&&0!==e.tag&&Ss(),o=e.pendingLanes,0!==(1&o)?e===Gi?Xi++:(Xi=0,Gi=e):Xi=0,Hl()}(e,n,t,r)}finally{Ni.transition=l,kn=r}return null}function Ss(){if(null!==Ki){var e=wn(Yi),n=Ni.transition,t=kn;try{if(Ni.transition=null,kn=16>e?16:e,null===Ki)var r=!1;else{if(e=Ki,Ki=null,Yi=0,0!==(6&Pi))throw Error(a(331));var l=Pi;for(Pi|=4,Ju=e.current;null!==Ju;){var o=Ju,u=o.child;if(0!==(16&Ju.flags)){var i=o.deletions;if(null!==i){for(var s=0;s<i.length;s++){var c=i[s];for(Ju=c;null!==Ju;){var f=Ju;switch(f.tag){case 0:case 11:case 15:ti(8,f,o)}var d=f.child;if(null!==d)d.return=f,Ju=d;else for(;null!==Ju;){var p=(f=Ju).sibling,m=f.return;if(ai(f),f===c){Ju=null;break}if(null!==p){p.return=m,Ju=p;break}Ju=m}}}var h=o.alternate;if(null!==h){var v=h.child;if(null!==v){h.child=null;do{var g=v.sibling;v.sibling=null,v=g}while(null!==v)}}Ju=o}}if(0!==(2064&o.subtreeFlags)&&null!==u)u.return=o,Ju=u;else e:for(;null!==Ju;){if(0!==(2048&(o=Ju).flags))switch(o.tag){case 0:case 11:case 15:ti(9,o,o.return)}var y=o.sibling;if(null!==y){y.return=o.return,Ju=y;break e}Ju=o.return}}var b=e.current;for(Ju=b;null!==Ju;){var k=(u=Ju).child;if(0!==(2064&u.subtreeFlags)&&null!==k)k.return=u,Ju=k;else e:for(u=b;null!==Ju;){if(0!==(2048&(i=Ju).flags))try{switch(i.tag){case 0:case 11:case 15:ri(9,i)}}catch(S){Es(i,i.return,S)}if(i===u){Ju=null;break e}var w=i.sibling;if(null!==w){w.return=i.return,Ju=w;break e}Ju=i.return}}if(Pi=l,Hl(),an&&"function"===typeof an.onPostCommitFiberRoot)try{an.onPostCommitFiberRoot(ln,e)}catch(S){}r=!0}return r}finally{kn=t,Ni.transition=n}}return!1}function xs(e,n,t){e=Ua(e,n=pu(0,n=su(t,n),1),1),n=es(),null!==e&&(yn(e,1,n),rs(e,n))}function Es(e,n,t){if(3===e.tag)xs(e,e,t);else for(;null!==n;){if(3===n.tag){xs(n,e,t);break}if(1===n.tag){var r=n.stateNode;if("function"===typeof n.type.getDerivedStateFromError||"function"===typeof r.componentDidCatch&&(null===Qi||!Qi.has(r))){n=Ua(n,e=mu(n,e=su(t,e),1),1),e=es(),null!==n&&(yn(n,1,e),rs(n,e));break}}n=n.return}}function Cs(e,n,t){var r=e.pingCache;null!==r&&r.delete(n),n=es(),e.pingedLanes|=e.suspendedLanes&t,zi===e&&(Ti&t)===t&&(4===Mi||3===Mi&&(130023424&Ti)===Ti&&500>Ge()-Ai?ds(e,0):Di|=t),rs(e,n)}function _s(e,n){0===n&&(0===(1&e.mode)?n=1:(n=fn,0===(130023424&(fn<<=1))&&(fn=4194304)));var t=es();null!==(e=Ma(e,n))&&(yn(e,n,t),rs(e,t))}function Ns(e){var n=e.memoizedState,t=0;null!==n&&(t=n.retryLane),_s(e,t)}function Ps(e,n){var t=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;null!==l&&(t=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(a(314))}null!==r&&r.delete(n),_s(e,t)}function zs(e,n){return qe(e,n)}function Ls(e,n,t,r){this.tag=e,this.key=t,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=n,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Ts(e,n,t,r){return new Ls(e,n,t,r)}function Rs(e){return!(!(e=e.prototype)||!e.isReactComponent)}function Os(e,n){var t=e.alternate;return null===t?((t=Ts(e.tag,n,e.key,e.mode)).elementType=e.elementType,t.type=e.type,t.stateNode=e.stateNode,t.alternate=e,e.alternate=t):(t.pendingProps=n,t.type=e.type,t.flags=0,t.subtreeFlags=0,t.deletions=null),t.flags=14680064&e.flags,t.childLanes=e.childLanes,t.lanes=e.lanes,t.child=e.child,t.memoizedProps=e.memoizedProps,t.memoizedState=e.memoizedState,t.updateQueue=e.updateQueue,n=e.dependencies,t.dependencies=null===n?null:{lanes:n.lanes,firstContext:n.firstContext},t.sibling=e.sibling,t.index=e.index,t.ref=e.ref,t}function Ms(e,n,t,r,l,o){var u=2;if(r=e,"function"===typeof e)Rs(e)&&(u=1);else if("string"===typeof e)u=5;else e:switch(e){case x:return js(t.children,l,o,n);case E:u=8,l|=8;break;case C:return(e=Ts(12,t,n,2|l)).elementType=C,e.lanes=o,e;case z:return(e=Ts(13,t,n,l)).elementType=z,e.lanes=o,e;case L:return(e=Ts(19,t,n,l)).elementType=L,e.lanes=o,e;case O:return Fs(t,l,o,n);default:if("object"===typeof e&&null!==e)switch(e.$$typeof){case _:u=10;break e;case N:u=9;break e;case P:u=11;break e;case T:u=14;break e;case R:u=16,r=null;break e}throw Error(a(130,null==e?e:typeof e,""))}return(n=Ts(u,t,n,l)).elementType=e,n.type=r,n.lanes=o,n}function js(e,n,t,r){return(e=Ts(7,e,r,n)).lanes=t,e}function Fs(e,n,t,r){return(e=Ts(22,e,r,n)).elementType=O,e.lanes=t,e.stateNode={isHidden:!1},e}function Is(e,n,t){return(e=Ts(6,e,null,n)).lanes=t,e}function Ds(e,n,t){return(n=Ts(4,null!==e.children?e.children:[],e.key,n)).lanes=t,n.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},n}function Us(e,n,t,r,l){this.tag=n,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=gn(0),this.expirationTimes=gn(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=gn(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function $s(e,n,t,r,l,a,o,u,i){return e=new Us(e,n,t,u,i),1===n?(n=1,!0===a&&(n|=8)):n=0,a=Ts(3,null,null,n),e.current=a,a.stateNode=e,a.memoizedState={element:r,isDehydrated:t,cache:null,transitions:null,pendingSuspenseBoundaries:null},Fa(a),e}function As(e){if(!e)return Nl;e:{if(He(e=e._reactInternals)!==e||1!==e.tag)throw Error(a(170));var n=e;do{switch(n.tag){case 3:n=n.stateNode.context;break e;case 1:if(Rl(n.type)){n=n.stateNode.__reactInternalMemoizedMergedChildContext;break e}}n=n.return}while(null!==n);throw Error(a(171))}if(1===e.tag){var t=e.type;if(Rl(t))return jl(e,t,n)}return n}function Hs(e,n,t,r,l,a,o,u,i){return(e=$s(t,r,!0,e,0,a,0,u,i)).context=As(null),t=e.current,(a=Da(r=es(),l=ns(t))).callback=void 0!==n&&null!==n?n:null,Ua(t,a,l),e.current.lanes=l,yn(e,l,r),rs(e,r),e}function Vs(e,n,t,r){var l=n.current,a=es(),o=ns(l);return t=As(t),null===n.context?n.context=t:n.pendingContext=t,(n=Da(a,o)).payload={element:e},null!==(r=void 0===r?null:r)&&(n.callback=r),null!==(e=Ua(l,n,o))&&(ts(e,l,o,a),$a(e,l,o)),o}function Bs(e){return(e=e.current).child?(e.child.tag,e.child.stateNode):null}function Ws(e,n){if(null!==(e=e.memoizedState)&&null!==e.dehydrated){var t=e.retryLane;e.retryLane=0!==t&&t<n?t:n}}function Qs(e,n){Ws(e,n),(e=e.alternate)&&Ws(e,n)}xi=function(e,n,t){if(null!==e)if(e.memoizedProps!==n.pendingProps||zl.current)bu=!0;else{if(0===(e.lanes&t)&&0===(128&n.flags))return bu=!1,function(e,n,t){switch(n.tag){case 3:zu(n),pa();break;case 5:Ga(n);break;case 1:Rl(n.type)&&Fl(n);break;case 4:Ya(n,n.stateNode.containerInfo);break;case 10:var r=n.type._context,l=n.memoizedProps.value;_l(Sa,r._currentValue),r._currentValue=l;break;case 13:if(null!==(r=n.memoizedState))return null!==r.dehydrated?(_l(Za,1&Za.current),n.flags|=128,null):0!==(t&n.child.childLanes)?Iu(e,n,t):(_l(Za,1&Za.current),null!==(e=Bu(e,n,t))?e.sibling:null);_l(Za,1&Za.current);break;case 19:if(r=0!==(t&n.childLanes),0!==(128&e.flags)){if(r)return Hu(e,n,t);n.flags|=128}if(null!==(l=n.memoizedState)&&(l.rendering=null,l.tail=null,l.lastEffect=null),_l(Za,Za.current),r)break;return null;case 22:case 23:return n.lanes=0,Eu(e,n,t)}return Bu(e,n,t)}(e,n,t);bu=0!==(131072&e.flags)}else bu=!1,la&&0!==(1048576&n.flags)&&Zl(n,Ql,n.index);switch(n.lanes=0,n.tag){case 2:var r=n.type;Vu(e,n),e=n.pendingProps;var l=Tl(n,Pl.current);za(n,t),l=vo(null,n,r,e,l,t);var o=go();return n.flags|=1,"object"===typeof l&&null!==l&&"function"===typeof l.render&&void 0===l.$$typeof?(n.tag=1,n.memoizedState=null,n.updateQueue=null,Rl(r)?(o=!0,Fl(n)):o=!1,n.memoizedState=null!==l.state&&void 0!==l.state?l.state:null,Fa(n),l.updater=lu,n.stateNode=l,l._reactInternals=n,iu(n,r,e,t),n=Pu(null,n,r,!0,o,t)):(n.tag=0,la&&o&&ea(n),ku(null,n,l,t),n=n.child),n;case 16:r=n.elementType;e:{switch(Vu(e,n),e=n.pendingProps,r=(l=r._init)(r._payload),n.type=r,l=n.tag=function(e){if("function"===typeof e)return Rs(e)?1:0;if(void 0!==e&&null!==e){if((e=e.$$typeof)===P)return 11;if(e===T)return 14}return 2}(r),e=tu(r,e),l){case 0:n=_u(null,n,r,e,t);break e;case 1:n=Nu(null,n,r,e,t);break e;case 11:n=wu(null,n,r,e,t);break e;case 14:n=Su(null,n,r,tu(r.type,e),t);break e}throw Error(a(306,r,""))}return n;case 0:return r=n.type,l=n.pendingProps,_u(e,n,r,l=n.elementType===r?l:tu(r,l),t);case 1:return r=n.type,l=n.pendingProps,Nu(e,n,r,l=n.elementType===r?l:tu(r,l),t);case 3:e:{if(zu(n),null===e)throw Error(a(387));r=n.pendingProps,l=(o=n.memoizedState).element,Ia(e,n),Ha(n,r,null,t);var u=n.memoizedState;if(r=u.element,o.isDehydrated){if(o={element:r,isDehydrated:!1,cache:u.cache,pendingSuspenseBoundaries:u.pendingSuspenseBoundaries,transitions:u.transitions},n.updateQueue.baseState=o,n.memoizedState=o,256&n.flags){n=Lu(e,n,r,t,l=su(Error(a(423)),n));break e}if(r!==l){n=Lu(e,n,r,t,l=su(Error(a(424)),n));break e}for(ra=sl(n.stateNode.containerInfo.firstChild),ta=n,la=!0,aa=null,t=wa(n,null,r,t),n.child=t;t;)t.flags=-3&t.flags|4096,t=t.sibling}else{if(pa(),r===l){n=Bu(e,n,t);break e}ku(e,n,r,t)}n=n.child}return n;case 5:return Ga(n),null===e&&sa(n),r=n.type,l=n.pendingProps,o=null!==e?e.memoizedProps:null,u=l.children,tl(r,l)?u=null:null!==o&&tl(r,o)&&(n.flags|=32),Cu(e,n),ku(e,n,u,t),n.child;case 6:return null===e&&sa(n),null;case 13:return Iu(e,n,t);case 4:return Ya(n,n.stateNode.containerInfo),r=n.pendingProps,null===e?n.child=ka(n,null,r,t):ku(e,n,r,t),n.child;case 11:return r=n.type,l=n.pendingProps,wu(e,n,r,l=n.elementType===r?l:tu(r,l),t);case 7:return ku(e,n,n.pendingProps,t),n.child;case 8:case 12:return ku(e,n,n.pendingProps.children,t),n.child;case 10:e:{if(r=n.type._context,l=n.pendingProps,o=n.memoizedProps,u=l.value,_l(Sa,r._currentValue),r._currentValue=u,null!==o)if(ur(o.value,u)){if(o.children===l.children&&!zl.current){n=Bu(e,n,t);break e}}else for(null!==(o=n.child)&&(o.return=n);null!==o;){var i=o.dependencies;if(null!==i){u=o.child;for(var s=i.firstContext;null!==s;){if(s.context===r){if(1===o.tag){(s=Da(-1,t&-t)).tag=2;var c=o.updateQueue;if(null!==c){var f=(c=c.shared).pending;null===f?s.next=s:(s.next=f.next,f.next=s),c.pending=s}}o.lanes|=t,null!==(s=o.alternate)&&(s.lanes|=t),Pa(o.return,t,n),i.lanes|=t;break}s=s.next}}else if(10===o.tag)u=o.type===n.type?null:o.child;else if(18===o.tag){if(null===(u=o.return))throw Error(a(341));u.lanes|=t,null!==(i=u.alternate)&&(i.lanes|=t),Pa(u,t,n),u=o.sibling}else u=o.child;if(null!==u)u.return=o;else for(u=o;null!==u;){if(u===n){u=null;break}if(null!==(o=u.sibling)){o.return=u.return,u=o;break}u=u.return}o=u}ku(e,n,l.children,t),n=n.child}return n;case 9:return l=n.type,r=n.pendingProps.children,za(n,t),r=r(l=La(l)),n.flags|=1,ku(e,n,r,t),n.child;case 14:return l=tu(r=n.type,n.pendingProps),Su(e,n,r,l=tu(r.type,l),t);case 15:return xu(e,n,n.type,n.pendingProps,t);case 17:return r=n.type,l=n.pendingProps,l=n.elementType===r?l:tu(r,l),Vu(e,n),n.tag=1,Rl(r)?(e=!0,Fl(n)):e=!1,za(n,t),ou(n,r,l),iu(n,r,l,t),Pu(null,n,r,!0,e,t);case 19:return Hu(e,n,t);case 22:return Eu(e,n,t)}throw Error(a(156,n.tag))};var qs="function"===typeof reportError?reportError:function(e){console.error(e)};function Ks(e){this._internalRoot=e}function Ys(e){this._internalRoot=e}function Xs(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType)}function Gs(e){return!(!e||1!==e.nodeType&&9!==e.nodeType&&11!==e.nodeType&&(8!==e.nodeType||" react-mount-point-unstable "!==e.nodeValue))}function Js(){}function Zs(e,n,t,r,l){var a=t._reactRootContainer;if(a){var o=a;if("function"===typeof l){var u=l;l=function(){var e=Bs(o);u.call(e)}}Vs(n,o,e,l)}else o=function(e,n,t,r,l){if(l){if("function"===typeof r){var a=r;r=function(){var e=Bs(o);a.call(e)}}var o=Hs(n,r,e,0,null,!1,0,"",Js);return e._reactRootContainer=o,e[ml]=o.current,Hr(8===e.nodeType?e.parentNode:e),cs(),o}for(;l=e.lastChild;)e.removeChild(l);if("function"===typeof r){var u=r;r=function(){var e=Bs(i);u.call(e)}}var i=$s(e,0,!1,null,0,!1,0,"",Js);return e._reactRootContainer=i,e[ml]=i.current,Hr(8===e.nodeType?e.parentNode:e),cs((function(){Vs(n,i,t,r)})),i}(t,n,e,l,r);return Bs(o)}Ys.prototype.render=Ks.prototype.render=function(e){var n=this._internalRoot;if(null===n)throw Error(a(409));Vs(e,n,null,null)},Ys.prototype.unmount=Ks.prototype.unmount=function(){var e=this._internalRoot;if(null!==e){this._internalRoot=null;var n=e.containerInfo;cs((function(){Vs(null,e,null,null)})),n[ml]=null}},Ys.prototype.unstable_scheduleHydration=function(e){if(e){var n=Cn();e={blockedOn:null,target:e,priority:n};for(var t=0;t<Mn.length&&0!==n&&n<Mn[t].priority;t++);Mn.splice(t,0,e),0===t&&Dn(e)}},Sn=function(e){switch(e.tag){case 3:var n=e.stateNode;if(n.current.memoizedState.isDehydrated){var t=dn(n.pendingLanes);0!==t&&(bn(n,1|t),rs(n,Ge()),0===(6&Pi)&&(Hi=Ge()+500,Hl()))}break;case 13:cs((function(){var n=Ma(e,1);if(null!==n){var t=es();ts(n,e,1,t)}})),Qs(e,1)}},xn=function(e){if(13===e.tag){var n=Ma(e,134217728);if(null!==n)ts(n,e,134217728,es());Qs(e,134217728)}},En=function(e){if(13===e.tag){var n=ns(e),t=Ma(e,n);if(null!==t)ts(t,e,n,es());Qs(e,n)}},Cn=function(){return kn},_n=function(e,n){var t=kn;try{return kn=e,n()}finally{kn=t}},Se=function(e,n,t){switch(n){case"input":if(J(e,t),n=t.name,"radio"===t.type&&null!=n){for(t=e;t.parentNode;)t=t.parentNode;for(t=t.querySelectorAll("input[name="+JSON.stringify(""+n)+'][type="radio"]'),n=0;n<t.length;n++){var r=t[n];if(r!==e&&r.form===e.form){var l=wl(r);if(!l)throw Error(a(90));q(r),J(r,l)}}}break;case"textarea":ae(e,t);break;case"select":null!=(n=t.value)&&te(e,!!t.multiple,n,!1)}},Pe=ss,ze=cs;var ec={usingClientEntryPoint:!1,Events:[bl,kl,wl,_e,Ne,ss]},nc={findFiberByHostInstance:yl,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},tc={bundleType:nc.bundleType,version:nc.version,rendererPackageName:nc.rendererPackageName,rendererConfig:nc.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:k.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return null===(e=We(e))?null:e.stateNode},findFiberByHostInstance:nc.findFiberByHostInstance||function(){return null},findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__){var rc=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!rc.isDisabled&&rc.supportsFiber)try{ln=rc.inject(tc),an=rc}catch(ce){}}n.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=ec,n.createPortal=function(e,n){var t=2<arguments.length&&void 0!==arguments[2]?arguments[2]:null;if(!Xs(n))throw Error(a(200));return function(e,n,t){var r=3<arguments.length&&void 0!==arguments[3]?arguments[3]:null;return{$$typeof:S,key:null==r?null:""+r,children:e,containerInfo:n,implementation:t}}(e,n,null,t)},n.createRoot=function(e,n){if(!Xs(e))throw Error(a(299));var t=!1,r="",l=qs;return null!==n&&void 0!==n&&(!0===n.unstable_strictMode&&(t=!0),void 0!==n.identifierPrefix&&(r=n.identifierPrefix),void 0!==n.onRecoverableError&&(l=n.onRecoverableError)),n=$s(e,1,!1,null,0,t,0,r,l),e[ml]=n.current,Hr(8===e.nodeType?e.parentNode:e),new Ks(n)},n.findDOMNode=function(e){if(null==e)return null;if(1===e.nodeType)return e;var n=e._reactInternals;if(void 0===n){if("function"===typeof e.render)throw Error(a(188));throw e=Object.keys(e).join(","),Error(a(268,e))}return e=null===(e=We(n))?null:e.stateNode},n.flushSync=function(e){return cs(e)},n.hydrate=function(e,n,t){if(!Gs(n))throw Error(a(200));return Zs(null,e,n,!0,t)},n.hydrateRoot=function(e,n,t){if(!Xs(e))throw Error(a(405));var r=null!=t&&t.hydratedSources||null,l=!1,o="",u=qs;if(null!==t&&void 0!==t&&(!0===t.unstable_strictMode&&(l=!0),void 0!==t.identifierPrefix&&(o=t.identifierPrefix),void 0!==t.onRecoverableError&&(u=t.onRecoverableError)),n=Hs(n,null,e,1,null!=t?t:null,l,0,o,u),e[ml]=n.current,Hr(e),r)for(e=0;e<r.length;e++)l=(l=(t=r[e])._getVersion)(t._source),null==n.mutableSourceEagerHydrationData?n.mutableSourceEagerHydrationData=[t,l]:n.mutableSourceEagerHydrationData.push(t,l);return new Ys(n)},n.render=function(e,n,t){if(!Gs(n))throw Error(a(200));return Zs(null,e,n,!1,t)},n.unmountComponentAtNode=function(e){if(!Gs(e))throw Error(a(40));return!!e._reactRootContainer&&(cs((function(){Zs(null,null,e,!1,(function(){e._reactRootContainer=null,e[ml]=null}))})),!0)},n.unstable_batchedUpdates=ss,n.unstable_renderSubtreeIntoContainer=function(e,n,t,r){if(!Gs(t))throw Error(a(200));if(null==e||void 0===e._reactInternals)throw Error(a(38));return Zs(e,n,t,!1,r)},n.version="18.3.1-next-f1338f8080-20240426"},391:(e,n,t)=>{var r=t(950);n.createRoot=r.createRoot,n.hydrateRoot=r.hydrateRoot},950:(e,n,t)=>{!function e(){if("undefined"!==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__&&"function"===typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE)try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(e)}catch(n){console.error(n)}}(),e.exports=t(730)},153:(e,n,t)=>{var r=t(43),l=Symbol.for("react.element"),a=Symbol.for("react.fragment"),o=Object.prototype.hasOwnProperty,u=r.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,i={key:!0,ref:!0,__self:!0,__source:!0};function s(e,n,t){var r,a={},s=null,c=null;for(r in void 0!==t&&(s=""+t),void 0!==n.key&&(s=""+n.key),void 0!==n.ref&&(c=n.ref),n)o.call(n,r)&&!i.hasOwnProperty(r)&&(a[r]=n[r]);if(e&&e.defaultProps)for(r in n=e.defaultProps)void 0===a[r]&&(a[r]=n[r]);return{$$typeof:l,type:e,key:s,ref:c,props:a,_owner:u.current}}n.Fragment=a,n.jsx=s,n.jsxs=s},202:(e,n)=>{var t=Symbol.for("react.element"),r=Symbol.for("react.portal"),l=Symbol.for("react.fragment"),a=Symbol.for("react.strict_mode"),o=Symbol.for("react.profiler"),u=Symbol.for("react.provider"),i=Symbol.for("react.context"),s=Symbol.for("react.forward_ref"),c=Symbol.for("react.suspense"),f=Symbol.for("react.memo"),d=Symbol.for("react.lazy"),p=Symbol.iterator;var m={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},h=Object.assign,v={};function g(e,n,t){this.props=e,this.context=n,this.refs=v,this.updater=t||m}function y(){}function b(e,n,t){this.props=e,this.context=n,this.refs=v,this.updater=t||m}g.prototype.isReactComponent={},g.prototype.setState=function(e,n){if("object"!==typeof e&&"function"!==typeof e&&null!=e)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,n,"setState")},g.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")},y.prototype=g.prototype;var k=b.prototype=new y;k.constructor=b,h(k,g.prototype),k.isPureReactComponent=!0;var w=Array.isArray,S=Object.prototype.hasOwnProperty,x={current:null},E={key:!0,ref:!0,__self:!0,__source:!0};function C(e,n,r){var l,a={},o=null,u=null;if(null!=n)for(l in void 0!==n.ref&&(u=n.ref),void 0!==n.key&&(o=""+n.key),n)S.call(n,l)&&!E.hasOwnProperty(l)&&(a[l]=n[l]);var i=arguments.length-2;if(1===i)a.children=r;else if(1<i){for(var s=Array(i),c=0;c<i;c++)s[c]=arguments[c+2];a.children=s}if(e&&e.defaultProps)for(l in i=e.defaultProps)void 0===a[l]&&(a[l]=i[l]);return{$$typeof:t,type:e,key:o,ref:u,props:a,_owner:x.current}}function _(e){return"object"===typeof e&&null!==e&&e.$$typeof===t}var N=/\/+/g;function P(e,n){return"object"===typeof e&&null!==e&&null!=e.key?function(e){var n={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,(function(e){return n[e]}))}(""+e.key):n.toString(36)}function z(e,n,l,a,o){var u=typeof e;"undefined"!==u&&"boolean"!==u||(e=null);var i=!1;if(null===e)i=!0;else switch(u){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case t:case r:i=!0}}if(i)return o=o(i=e),e=""===a?"."+P(i,0):a,w(o)?(l="",null!=e&&(l=e.replace(N,"$&/")+"/"),z(o,n,l,"",(function(e){return e}))):null!=o&&(_(o)&&(o=function(e,n){return{$$typeof:t,type:e.type,key:n,ref:e.ref,props:e.props,_owner:e._owner}}(o,l+(!o.key||i&&i.key===o.key?"":(""+o.key).replace(N,"$&/")+"/")+e)),n.push(o)),1;if(i=0,a=""===a?".":a+":",w(e))for(var s=0;s<e.length;s++){var c=a+P(u=e[s],s);i+=z(u,n,l,c,o)}else if(c=function(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=p&&e[p]||e["@@iterator"])?e:null}(e),"function"===typeof c)for(e=c.call(e),s=0;!(u=e.next()).done;)i+=z(u=u.value,n,l,c=a+P(u,s++),o);else if("object"===u)throw n=String(e),Error("Objects are not valid as a React child (found: "+("[object Object]"===n?"object with keys {"+Object.keys(e).join(", ")+"}":n)+"). If you meant to render a collection of children, use an array instead.");return i}function L(e,n,t){if(null==e)return e;var r=[],l=0;return z(e,r,"","",(function(e){return n.call(t,e,l++)})),r}function T(e){if(-1===e._status){var n=e._result;(n=n()).then((function(n){0!==e._status&&-1!==e._status||(e._status=1,e._result=n)}),(function(n){0!==e._status&&-1!==e._status||(e._status=2,e._result=n)})),-1===e._status&&(e._status=0,e._result=n)}if(1===e._status)return e._result.default;throw e._result}var R={current:null},O={transition:null},M={ReactCurrentDispatcher:R,ReactCurrentBatchConfig:O,ReactCurrentOwner:x};function j(){throw Error("act(...) is not supported in production builds of React.")}n.Children={map:L,forEach:function(e,n,t){L(e,(function(){n.apply(this,arguments)}),t)},count:function(e){var n=0;return L(e,(function(){n++})),n},toArray:function(e){return L(e,(function(e){return e}))||[]},only:function(e){if(!_(e))throw Error("React.Children.only expected to receive a single React element child.");return e}},n.Component=g,n.Fragment=l,n.Profiler=o,n.PureComponent=b,n.StrictMode=a,n.Suspense=c,n.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=M,n.act=j,n.cloneElement=function(e,n,r){if(null===e||void 0===e)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var l=h({},e.props),a=e.key,o=e.ref,u=e._owner;if(null!=n){if(void 0!==n.ref&&(o=n.ref,u=x.current),void 0!==n.key&&(a=""+n.key),e.type&&e.type.defaultProps)var i=e.type.defaultProps;for(s in n)S.call(n,s)&&!E.hasOwnProperty(s)&&(l[s]=void 0===n[s]&&void 0!==i?i[s]:n[s])}var s=arguments.length-2;if(1===s)l.children=r;else if(1<s){i=Array(s);for(var c=0;c<s;c++)i[c]=arguments[c+2];l.children=i}return{$$typeof:t,type:e.type,key:a,ref:o,props:l,_owner:u}},n.createContext=function(e){return(e={$$typeof:i,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null}).Provider={$$typeof:u,_context:e},e.Consumer=e},n.createElement=C,n.createFactory=function(e){var n=C.bind(null,e);return n.type=e,n},n.createRef=function(){return{current:null}},n.forwardRef=function(e){return{$$typeof:s,render:e}},n.isValidElement=_,n.lazy=function(e){return{$$typeof:d,_payload:{_status:-1,_result:e},_init:T}},n.memo=function(e,n){return{$$typeof:f,type:e,compare:void 0===n?null:n}},n.startTransition=function(e){var n=O.transition;O.transition={};try{e()}finally{O.transition=n}},n.unstable_act=j,n.useCallback=function(e,n){return R.current.useCallback(e,n)},n.useContext=function(e){return R.current.useContext(e)},n.useDebugValue=function(){},n.useDeferredValue=function(e){return R.current.useDeferredValue(e)},n.useEffect=function(e,n){return R.current.useEffect(e,n)},n.useId=function(){return R.current.useId()},n.useImperativeHandle=function(e,n,t){return R.current.useImperativeHandle(e,n,t)},n.useInsertionEffect=function(e,n){return R.current.useInsertionEffect(e,n)},n.useLayoutEffect=function(e,n){return R.current.useLayoutEffect(e,n)},n.useMemo=function(e,n){return R.current.useMemo(e,n)},n.useReducer=function(e,n,t){return R.current.useReducer(e,n,t)},n.useRef=function(e){return R.current.useRef(e)},n.useState=function(e){return R.current.useState(e)},n.useSyncExternalStore=function(e,n,t){return R.current.useSyncExternalStore(e,n,t)},n.useTransition=function(){return R.current.useTransition()},n.version="18.3.1"},43:(e,n,t)=>{e.exports=t(202)},579:(e,n,t)=>{e.exports=t(153)},234:(e,n)=>{function t(e,n){var t=e.length;e.push(n);e:for(;0<t;){var r=t-1>>>1,l=e[r];if(!(0<a(l,n)))break e;e[r]=n,e[t]=l,t=r}}function r(e){return 0===e.length?null:e[0]}function l(e){if(0===e.length)return null;var n=e[0],t=e.pop();if(t!==n){e[0]=t;e:for(var r=0,l=e.length,o=l>>>1;r<o;){var u=2*(r+1)-1,i=e[u],s=u+1,c=e[s];if(0>a(i,t))s<l&&0>a(c,i)?(e[r]=c,e[s]=t,r=s):(e[r]=i,e[u]=t,r=u);else{if(!(s<l&&0>a(c,t)))break e;e[r]=c,e[s]=t,r=s}}}return n}function a(e,n){var t=e.sortIndex-n.sortIndex;return 0!==t?t:e.id-n.id}if("object"===typeof performance&&"function"===typeof performance.now){var o=performance;n.unstable_now=function(){return o.now()}}else{var u=Date,i=u.now();n.unstable_now=function(){return u.now()-i}}var s=[],c=[],f=1,d=null,p=3,m=!1,h=!1,v=!1,g="function"===typeof setTimeout?setTimeout:null,y="function"===typeof clearTimeout?clearTimeout:null,b="undefined"!==typeof setImmediate?setImmediate:null;function k(e){for(var n=r(c);null!==n;){if(null===n.callback)l(c);else{if(!(n.startTime<=e))break;l(c),n.sortIndex=n.expirationTime,t(s,n)}n=r(c)}}function w(e){if(v=!1,k(e),!h)if(null!==r(s))h=!0,O(S);else{var n=r(c);null!==n&&M(w,n.startTime-e)}}function S(e,t){h=!1,v&&(v=!1,y(_),_=-1),m=!0;var a=p;try{for(k(t),d=r(s);null!==d&&(!(d.expirationTime>t)||e&&!z());){var o=d.callback;if("function"===typeof o){d.callback=null,p=d.priorityLevel;var u=o(d.expirationTime<=t);t=n.unstable_now(),"function"===typeof u?d.callback=u:d===r(s)&&l(s),k(t)}else l(s);d=r(s)}if(null!==d)var i=!0;else{var f=r(c);null!==f&&M(w,f.startTime-t),i=!1}return i}finally{d=null,p=a,m=!1}}"undefined"!==typeof navigator&&void 0!==navigator.scheduling&&void 0!==navigator.scheduling.isInputPending&&navigator.scheduling.isInputPending.bind(navigator.scheduling);var x,E=!1,C=null,_=-1,N=5,P=-1;function z(){return!(n.unstable_now()-P<N)}function L(){if(null!==C){var e=n.unstable_now();P=e;var t=!0;try{t=C(!0,e)}finally{t?x():(E=!1,C=null)}}else E=!1}if("function"===typeof b)x=function(){b(L)};else if("undefined"!==typeof MessageChannel){var T=new MessageChannel,R=T.port2;T.port1.onmessage=L,x=function(){R.postMessage(null)}}else x=function(){g(L,0)};function O(e){C=e,E||(E=!0,x())}function M(e,t){_=g((function(){e(n.unstable_now())}),t)}n.unstable_IdlePriority=5,n.unstable_ImmediatePriority=1,n.unstable_LowPriority=4,n.unstable_NormalPriority=3,n.unstable_Profiling=null,n.unstable_UserBlockingPriority=2,n.unstable_cancelCallback=function(e){e.callback=null},n.unstable_continueExecution=function(){h||m||(h=!0,O(S))},n.unstable_forceFrameRate=function(e){0>e||125<e?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):N=0<e?Math.floor(1e3/e):5},n.unstable_getCurrentPriorityLevel=function(){return p},n.unstable_getFirstCallbackNode=function(){return r(s)},n.unstable_next=function(e){switch(p){case 1:case 2:case 3:var n=3;break;default:n=p}var t=p;p=n;try{return e()}finally{p=t}},n.unstable_pauseExecution=function(){},n.unstable_requestPaint=function(){},n.unstable_runWithPriority=function(e,n){switch(e){case 1:case 2:case 3:case 4:case 5:break;default:e=3}var t=p;p=e;try{return n()}finally{p=t}},n.unstable_scheduleCallback=function(e,l,a){var o=n.unstable_now();switch("object"===typeof a&&null!==a?a="number"===typeof(a=a.delay)&&0<a?o+a:o:a=o,e){case 1:var u=-1;break;case 2:u=250;break;case 5:u=1073741823;break;case 4:u=1e4;break;default:u=5e3}return e={id:f++,callback:l,priorityLevel:e,startTime:a,expirationTime:u=a+u,sortIndex:-1},a>o?(e.sortIndex=a,t(c,e),null===r(s)&&e===r(c)&&(v?(y(_),_=-1):v=!0,M(w,a-o))):(e.sortIndex=u,t(s,e),h||m||(h=!0,O(S))),e},n.unstable_shouldYield=z,n.unstable_wrapCallback=function(e){var n=p;return function(){var t=p;p=n;try{return e.apply(this,arguments)}finally{p=t}}}},853:(e,n,t)=>{e.exports=t(234)},540:(e,n)=>{var t=Symbol.for("react.element"),r=Symbol.for("react.portal"),l=Symbol.for("react.fragment"),a=Symbol.for("react.strict_mode"),o=Symbol.for("react.profiler"),u=Symbol.for("react.provider"),i=Symbol.for("react.context"),s=Symbol.for("react.forward_ref"),c=Symbol.for("react.suspense"),f=Symbol.for("react.memo"),d=Symbol.for("react.lazy"),p=Symbol.iterator;var m={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},h=Object.assign,v={};function g(e,n,t){this.props=e,this.context=n,this.refs=v,this.updater=t||m}function y(){}function b(e,n,t){this.props=e,this.context=n,this.refs=v,this.updater=t||m}g.prototype.isReactComponent={},g.prototype.setState=function(e,n){if("object"!==typeof e&&"function"!==typeof e&&null!=e)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,n,"setState")},g.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")},y.prototype=g.prototype;var k=b.prototype=new y;k.constructor=b,h(k,g.prototype),k.isPureReactComponent=!0;var w=Array.isArray,S=Object.prototype.hasOwnProperty,x={current:null},E={key:!0,ref:!0,__self:!0,__source:!0};function C(e,n,r){var l,a={},o=null,u=null;if(null!=n)for(l in void 0!==n.ref&&(u=n.ref),void 0!==n.key&&(o=""+n.key),n)S.call(n,l)&&!E.hasOwnProperty(l)&&(a[l]=n[l]);var i=arguments.length-2;if(1===i)a.children=r;else if(1<i){for(var s=Array(i),c=0;c<i;c++)s[c]=arguments[c+2];a.children=s}if(e&&e.defaultProps)for(l in i=e.defaultProps)void 0===a[l]&&(a[l]=i[l]);return{$$typeof:t,type:e,key:o,ref:u,props:a,_owner:x.current}}function _(e){return"object"===typeof e&&null!==e&&e.$$typeof===t}var N=/\/+/g;function P(e,n){return"object"===typeof e&&null!==e&&null!=e.key?function(e){var n={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,(function(e){return n[e]}))}(""+e.key):n.toString(36)}function z(e,n,l,a,o){var u=typeof e;"undefined"!==u&&"boolean"!==u||(e=null);var i=!1;if(null===e)i=!0;else switch(u){case"string":case"number":i=!0;break;case"object":switch(e.$$typeof){case t:case r:i=!0}}if(i)return o=o(i=e),e=""===a?"."+P(i,0):a,w(o)?(l="",null!=e&&(l=e.replace(N,"$&/")+"/"),z(o,n,l,"",(function(e){return e}))):null!=o&&(_(o)&&(o=function(e,n){return{$$typeof:t,type:e.type,key:n,ref:e.ref,props:e.props,_owner:e._owner}}(o,l+(!o.key||i&&i.key===o.key?"":(""+o.key).replace(N,"$&/")+"/")+e)),n.push(o)),1;if(i=0,a=""===a?".":a+":",w(e))for(var s=0;s<e.length;s++){var c=a+P(u=e[s],s);i+=z(u,n,l,c,o)}else if(c=function(e){return null===e||"object"!==typeof e?null:"function"===typeof(e=p&&e[p]||e["@@iterator"])?e:null}(e),"function"===typeof c)for(e=c.call(e),s=0;!(u=e.next()).done;)i+=z(u=u.value,n,l,c=a+P(u,s++),o);else if("object"===u)throw n=String(e),Error("Objects are not valid as a React child (found: "+("[object Object]"===n?"object with keys {"+Object.keys(e).join(", ")+"}":n)+"). If you meant to render a collection of children, use an array instead.");return i}function L(e,n,t){if(null==e)return e;var r=[],l=0;return z(e,r,"","",(function(e){return n.call(t,e,l++)})),r}function T(e){if(-1===e._status){var n=e._result;(n=n()).then((function(n){0!==e._status&&-1!==e._status||(e._status=1,e._result=n)}),(function(n){0!==e._status&&-1!==e._status||(e._status=2,e._result=n)})),-1===e._status&&(e._status=0,e._result=n)}if(1===e._status)return e._result.default;throw e._result}var R={current:null},O={transition:null};function M(){throw Error("act(...) is not supported in production builds of React.")}n.createElement=C,n.forwardRef=function(e){return{$$typeof:s,render:e}}},949:(e,n,t)=>{e.exports=t(540)}},n={};function t(r){var l=n[r];if(void 0!==l)return l.exports;var a=n[r]={exports:{}};return e[r](a,a.exports,t),a.exports}t.m=e,t.d=(e,n)=>{for(var r in n)t.o(n,r)&&!t.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:n[r]})},t.f={},t.e=e=>Promise.all(Object.keys(t.f).reduce(((n,r)=>(t.f[r](e,n),n)),[])),t.u=e=>"static/js/"+e+".23c8dce4.chunk.js",t.miniCssF=e=>{},t.o=(e,n)=>Object.prototype.hasOwnProperty.call(e,n),(()=>{var e={},n="nook-story:";t.l=(r,l,a,o)=>{if(e[r])e[r].push(l);else{var u,i;if(void 0!==a)for(var s=document.getElementsByTagName("script"),c=0;c<s.length;c++){var f=s[c];if(f.getAttribute("src")==r||f.getAttribute("data-webpack")==n+a){u=f;break}}u||(i=!0,(u=document.createElement("script")).charset="utf-8",u.timeout=120,t.nc&&u.setAttribute("nonce",t.nc),u.setAttribute("data-webpack",n+a),u.src=r),e[r]=[l];var d=(n,t)=>{u.onerror=u.onload=null,clearTimeout(p);var l=e[r];if(delete e[r],u.parentNode&&u.parentNode.removeChild(u),l&&l.forEach((e=>e(t))),n)return n(t)},p=setTimeout(d.bind(null,void 0,{type:"timeout",target:u}),12e4);u.onerror=d.bind(null,u.onerror),u.onload=d.bind(null,u.onload),i&&document.head.appendChild(u)}}})(),t.r=e=>{"undefined"!==typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},t.p="/",(()=>{var e={792:0};t.f.j=(n,r)=>{var l=t.o(e,n)?e[n]:void 0;if(0!==l)if(l)r.push(l[2]);else{var a=new Promise(((t,r)=>l=e[n]=[t,r]));r.push(l[2]=a);var o=t.p+t.u(n),u=new Error;t.l(o,(r=>{if(t.o(e,n)&&(0!==(l=e[n])&&(e[n]=void 0),l)){var a=r&&("load"===r.type?"missing":r.type),o=r&&r.target&&r.target.src;u.message="Loading chunk "+n+" failed.\n("+a+": "+o+")",u.name="ChunkLoadError",u.type=a,u.request=o,l[1](u)}}),"chunk-"+n,n)}};var n=(n,r)=>{var l,a,o=r[0],u=r[1],i=r[2],s=0;if(o.some((n=>0!==e[n]))){for(l in u)t.o(u,l)&&(t.m[l]=u[l]);if(i)i(t)}for(n&&n(r);s<o.length;s++)a=o[s],t.o(e,a)&&e[a]&&e[a][0](),e[a]=0},r=self.webpackChunknook_story=self.webpackChunknook_story||[];r.forEach(n.bind(null,0)),r.push=n.bind(null,r.push.bind(r))})();var r=t(43),l=t(391),a=t(949);const o=function(){for(var e=arguments.length,n=new Array(e),t=0;t<e;t++)n[t]=arguments[t];return n.filter(((e,n,t)=>Boolean(e)&&""!==e.trim()&&t.indexOf(e)===n)).join(" ").trim()};var u={xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:2,strokeLinecap:"round",strokeLinejoin:"round"};const i=(0,a.forwardRef)(((e,n)=>{let{color:t="currentColor",size:r=24,strokeWidth:l=2,absoluteStrokeWidth:i,className:s="",children:c,iconNode:f,...d}=e;return(0,a.createElement)("svg",{ref:n,...u,width:r,height:r,stroke:t,strokeWidth:i?24*Number(l)/Number(r):l,className:o("lucide",s),...d},[...f.map((e=>{let[n,t]=e;return(0,a.createElement)(n,t)})),...Array.isArray(c)?c:[c]])})),s=(e,n)=>{const t=(0,a.forwardRef)(((t,r)=>{let{className:l,...u}=t;return(0,a.createElement)(i,{ref:r,iconNode:n,className:o(`lucide-${s=e,s.replace(/([a-z0-9])([A-Z])/g,"$1-$2").toLowerCase()}`,l),...u});var s}));return t.displayName=`${e}`,t},c=s("Plus",[["path",{d:"M5 12h14",key:"1ays0h"}],["path",{d:"M12 5v14",key:"s699le"}]]),f=s("House",[["path",{d:"M15 21v-8a1 1 0 0 0-1-1h-4a1 1 0 0 0-1 1v8",key:"5wwlr5"}],["path",{d:"M3 10a2 2 0 0 1 .709-1.528l7-5.999a2 2 0 0 1 2.582 0l7 5.999A2 2 0 0 1 21 10v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z",key:"1d0kgt"}]]);var d=t(579);const p=e=>{let{onHouseSelected:n}=e;const[t,l]=(0,r.useState)([]),[a,o]=(0,r.useState)(!0);(0,r.useEffect)((()=>{const e=JSON.parse(localStorage.getItem("nookHouses")||"[]");l(e),o(0===e.length)}),[]);const u=()=>{const e={id:Date.now().toString(),name:"New NookHouse",primaryColor:"#6a5acd",createdAt:(new Date).toISOString()},r=[...t,e];localStorage.setItem("nookHouses",JSON.stringify(r)),l(r),n(e)};return(0,d.jsxs)("div",{className:"nookhouse-menu p-6 bg-purple-600 min-h-screen flex flex-col justify-center items-center",children:[(0,d.jsx)("h1",{className:"text-4xl font-bold text-white mb-8",children:"Nook&Story"}),a?(0,d.jsxs)("div",{className:"text-center",children:[(0,d.jsx)("p",{className:"text-white mb-6",children:"Welcome! Create your first NookHouse"}),(0,d.jsxs)("button",{onClick:u,className:"bubble-btn flex items-center justify-center gap-2",children:[(0,d.jsx)(c,{})," Create New NookHouse"]})]}):(0,d.jsxs)("div",{className:"w-full max-w-md",children:[(0,d.jsx)("h2",{className:"text-2xl text-white mb-4",children:"Your NookHouses"}),(0,d.jsxs)("div",{className:"space-y-4",children:[t.map((e=>(0,d.jsxs)("button",{onClick:()=>(e=>{n(e)})(e),className:"bubble-btn w-full flex items-center gap-4",style:{backgroundColor:e.primaryColor},children:[(0,d.jsx)(f,{}),(0,d.jsx)("span",{children:e.name})]},e.id))),(0,d.jsxs)("button",{onClick:u,className:"bubble-btn w-full flex items-center justify-center gap-2",children:[(0,d.jsx)(c,{})," Add New NookHouse"]})]})]})]})},m=s("Settings",[["path",{d:"M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 1 1.72v.51a2 2 0 0 1-1 1.74l-.15.09a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.39a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1-1-1.74v-.5a2 2 0 0 1 1-1.74l.15-.09a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z",key:"1qme2f"}],["circle",{cx:"12",cy:"12",r:"3",key:"1v7zrd"}]]),h=s("Play",[["polygon",{points:"6 3 20 12 6 21 6 3",key:"1oa8hb"}]]),v=s("Image",[["rect",{width:"18",height:"18",x:"3",y:"3",rx:"2",ry:"2",key:"1m3agn"}],["circle",{cx:"9",cy:"9",r:"2",key:"af1f0g"}],["path",{d:"m21 15-3.086-3.086a2 2 0 0 0-2.828 0L6 21",key:"1xmnt7"}]]),g=s("X",[["path",{d:"M18 6 6 18",key:"1bl5f8"}],["path",{d:"m6 6 12 12",key:"d8bk6v"}]]),y={PUBLIC:"Public",PRIVATE:"Private",FRIENDS_ONLY:"Friends Only"},b=e=>{let{house:n,onClose:t,onUpdateHouse:l}=e;const[a,o]=(0,r.useState)(n.primaryColor),[u,i]=(0,r.useState)(n.accessLevel||y.PUBLIC);return(0,d.jsx)("div",{className:"modal-overlay",children:(0,d.jsxs)("div",{className:"modal-content",children:[(0,d.jsxs)("div",{className:"flex justify-between items-center mb-6",children:[(0,d.jsx)("h2",{className:"text-2xl font-bold",children:"Room Settings"}),(0,d.jsx)("button",{onClick:t,className:"text-gray-500",children:(0,d.jsx)(g,{size:24})})]}),(0,d.jsxs)("div",{className:"mb-6",children:[(0,d.jsx)("label",{className:"block text-lg font-medium mb-2",children:"Primary Color"}),(0,d.jsx)("input",{type:"color",value:a,onChange:e=>o(e.target.value),className:"w-full h-12 rounded"})]}),(0,d.jsxs)("div",{className:"mb-6",children:[(0,d.jsx)("label",{className:"block text-lg font-medium mb-2",children:"Access Level"}),(0,d.jsx)("select",{value:u,onChange:e=>i(e.target.value),className:"w-full h-12 rounded border-gray-300 focus:ring focus:ring-purple-500 focus:border-purple-500",children:Object.values(y).map((e=>(0,d.jsx)("option",{value:e,children:e},e)))})]}),(0,d.jsxs)("div",{className:"flex justify-between gap-4",children:[(0,d.jsx)("button",{onClick:t,className:"flex-1 py-2 px-4 rounded-lg bg-gray-100 text-gray-900 font-medium",children:"Cancel"}),(0,d.jsx)("button",{onClick:()=>{const e={...n,primaryColor:a,accessLevel:u},r=JSON.parse(localStorage.getItem("nookHouses")||"[]").map((t=>t.id===n.id?e:t));localStorage.setItem("nookHouses",JSON.stringify(r)),l(e),t()},className:"flex-1 py-2 px-4 rounded-lg bg-purple-600 text-white font-medium",children:"Save Changes"})]})]})})},k=e=>{let{house:n}=e;const[t,l]=(0,r.useState)([{id:"living-room",name:"Living Room",type:"living-room"},{id:"study-room",name:"Study Room",type:"study-room"},{id:"gaming-room",name:"Gaming Room",type:"gaming-room"},{id:"personal-room",name:"Personal Room",type:"personal-room"}]),[a,o]=(0,r.useState)(t[0]),[u,i]=(0,r.useState)(!1);return(0,d.jsxs)("div",{className:"rooms-menu",children:[(0,d.jsxs)("div",{className:"top-menu-bar",children:[(0,d.jsxs)("div",{className:"house-name",children:[n.name," "]}),(0,d.jsx)("button",{onClick:()=>i(!0),className:"settings-button",children:(0,d.jsx)(m,{size:20})})]}),(0,d.jsx)("h2",{className:"room-title",children:a.name}),(0,d.jsx)("div",{className:"room-display",children:(0,d.jsx)("div",{className:`room ${a.type}`})}),(0,d.jsxs)("div",{className:"media-decor-buttons",children:[(0,d.jsx)("button",{children:(0,d.jsx)(h,{size:24})}),(0,d.jsx)("button",{children:(0,d.jsx)(v,{size:24})})]}),(0,d.jsxs)("div",{className:"room-navigation",children:[(0,d.jsx)("div",{className:"rooms-list",children:t.map((e=>(0,d.jsx)("button",{onClick:()=>o(e),className:"room-button "+(a.id===e.id?"active":""),children:e.name},e.id)))}),(0,d.jsxs)("button",{onClick:()=>{const e={id:`room-${Date.now()}`,name:`New Room ${t.length+1}`,type:"custom-room"};l([...t,e])},className:"add-room-button",children:[(0,d.jsx)(c,{size:20}),"Add Room"]})]}),u&&(0,d.jsx)(b,{house:n,onClose:()=>i(!1),onUpdateHouse:e=>{console.log("House updated:",e)}})]})},w=()=>{const[e,n]=(0,r.useState)("nookhouse"),[t,l]=(0,r.useState)(null),[a,o]=(0,r.useState)(!1);return(0,d.jsxs)("div",{className:"app-container",children:["nookhouse"===e&&(0,d.jsx)(p,{onHouseSelected:e=>{l(e),n("rooms")}}),"rooms"===e&&t&&(0,d.jsxs)(d.Fragment,{children:[(0,d.jsx)(k,{house:t,onSettingsOpen:()=>o(!0)}),a&&(0,d.jsx)(b,{house:t,onClose:()=>o(!1),onUpdateHouse:e=>{l(e)}})]})]})},S=e=>{e&&e instanceof Function&&t.e(453).then(t.bind(t,453)).then((n=>{let{getCLS:t,getFID:r,getFCP:l,getLCP:a,getTTFB:o}=n;t(e),r(e),l(e),a(e),o(e)}))};l.createRoot(document.getElementById("root")).render((0,d.jsx)(r.StrictMode,{children:(0,d.jsx)(w,{})})),S()})();
+//# sourceMappingURL=main.f9304b0b.js.map
